Plutus.V1.Ledger.Interval

Description

A type for intervals and associated functions.

Documentation

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data UpperBound a

The upper bound of an interval.

Constructors

UpperBound (Extended a) Closure

data LowerBound a

The lower bound of an interval.

Constructors

LowerBound (Extended a) Closure

data Extended a

A set extended with a positive and negative infinity.

Constructors

NegInf

Finite a

PosInf

type Closure = Bool

Whether a bound is inclusive or not.

member :: Ord a => a -> Interval a -> Bool

Check whether a value is in an interval.

interval :: a -> a -> Interval a

interval a b includes all values that are greater than or equal to a and smaller than or equal to b. Therefore it includes a and b.

from :: a -> Interval a

from a is an Interval that includes all values that are greater than or equal to a.

to :: a -> Interval a

to a is an Interval that includes all values that are smaller than or equal to a.

always :: Interval a

An Interval that covers every slot.

never :: Interval a

An Interval that is empty.

singleton :: a -> Interval a

hull :: Ord a => Interval a -> Interval a -> Interval a

'hull a b' is the smallest interval containing a and b.

intersection :: Ord a => Interval a -> Interval a -> Interval a

'intersection a b' is the largest interval that is contained in a and in b, if it exists.

overlaps :: (Enum a, Ord a) => Interval a -> Interval a -> Bool

Check whether two intervals overlap, that is, whether there is a value that is a member of both intervals.

contains :: Ord a => Interval a -> Interval a -> Bool

a contains b is true if the Interval b is entirely contained in a. That is, a contains b if for every entry s, if member s b then member s a.

isEmpty :: (Enum a, Ord a) => Interval a -> Bool

Check if an Interval is empty.

before :: Ord a => a -> Interval a -> Bool

Check if a value is earlier than the beginning of an Interval.

after :: Ord a => a -> Interval a -> Bool

Check if a value is later than the end of a Interval.

lowerBound :: a -> LowerBound a

upperBound :: a -> UpperBound a

strictLowerBound :: a -> LowerBound a

strictUpperBound :: a -> UpperBound a