Plutus.V1.Ledger.Scripts

Description

Functions for working with scripts on the ledger.

Scripts

newtype Script

A script on the chain. This is an opaque type as far as the chain is concerned.

Constructors

Script

Fields

unScript :: Program DeBruijn DefaultUni DefaultFun ()

scriptSize :: Script -> Integer

The size of a Script. No particular interpretation is given to this, other than that it is proportional to the serialized size of the script.

fromCompiledCode :: CompiledCode a -> Script

Turn a CompiledCode (usually produced by compile) into a Script for use with this package.

data ScriptError

Constructors

EvaluationError [Text] String
Expected behavior of the engine (e.g. user-provided error)

EvaluationException String String
Unexpected behavior of the engine (a bug)

evaluateScript :: forall m. MonadError ScriptError m => Script -> m (ExBudget, [Text])

Evaluate a script, returning the trace log.

runScript :: MonadError ScriptError m => Context -> Validator -> Datum -> Redeemer -> m (ExBudget, [Text])

Evaluate a Validator with its Context, Datum, and Redeemer, returning the log.

runMintingPolicyScript :: MonadError ScriptError m => Context -> MintingPolicy -> Redeemer -> m (ExBudget, [Text])

Evaluate a MintingPolicy with its Context and Redeemer, returning the log.

runStakeValidatorScript :: MonadError ScriptError m => Context -> StakeValidator -> Redeemer -> m (ExBudget, [Text])

Evaluate a StakeValidator with its Context and Redeemer, returning the log.

applyValidator :: Context -> Validator -> Datum -> Redeemer -> Script

Apply a Validator to its Context, Datum, and Redeemer.

applyMintingPolicyScript :: Context -> MintingPolicy -> Redeemer -> Script

Apply MintingPolicy to its Context and Redeemer.

applyStakeValidatorScript :: Context -> StakeValidator -> Redeemer -> Script

Apply StakeValidator to its Context and Redeemer.

applyArguments :: Script -> [Data] -> Script

Script wrappers

mkValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> ()) -> Validator

newtype Validator

Validator is a wrapper around Scripts which are used as validators in transaction outputs.

Constructors

Validator

Fields

getValidator :: Script

unValidatorScript :: Validator -> Script

newtype Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

Constructors

Redeemer

Fields

getRedeemer :: BuiltinData

newtype Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

Constructors

Datum

Fields

getDatum :: BuiltinData

mkMintingPolicyScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> MintingPolicy

newtype MintingPolicy

MintingPolicy is a wrapper around Scripts which are used as validators for minting constraints.

Constructors

MintingPolicy

Fields

getMintingPolicy :: Script

unMintingPolicyScript :: MintingPolicy -> Script

mkStakeValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> StakeValidator

newtype StakeValidator

StakeValidator is a wrapper around Scripts which are used as validators for withdrawals and stake address certificates.

Constructors

StakeValidator

Fields

getStakeValidator :: Script

unStakeValidatorScript :: StakeValidator -> Script

newtype Context

Information about the state of the blockchain and about the transaction that is currently being validated, represented as a value in Data.

Constructors

Context BuiltinData

Hashes

newtype DatumHash

Script runtime representation of a Digest SHA256.

Constructors

DatumHash BuiltinByteString

newtype RedeemerHash

Script runtime representation of a Digest SHA256.

Constructors

RedeemerHash BuiltinByteString

newtype ScriptHash

Script runtime representation of a Digest SHA256.

Constructors

ScriptHash

Fields

getScriptHash :: BuiltinByteString

newtype ValidatorHash

Script runtime representation of a Digest SHA256.

Constructors

ValidatorHash BuiltinByteString

newtype MintingPolicyHash

Script runtime representation of a Digest SHA256.

Constructors

MintingPolicyHash BuiltinByteString

newtype StakeValidatorHash

Script runtime representation of a Digest SHA256.

Constructors

StakeValidatorHash BuiltinByteString

Example scripts

unitRedeemer :: Redeemer

() as a redeemer.

unitDatum :: Datum

() as a datum.