Plutus.V1.Ledger.Tx

Transactions

newtype TxId

A transaction ID, using a SHA256 hash as the transaction id.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data ScriptTag

A tag indicating the type of script that we are pointing to.

Constructors

Spend

Mint

Cert

Reward

data RedeemerPtr

A redeemer pointer is a pair of a script type tag t and an index i, picking out the ith script of type t in the transaction.

Constructors

RedeemerPtr ScriptTag Integer

type Redeemers = Map RedeemerPtr Redeemer

Transaction outputs

data TxOut

A transaction output, consisting of a target address, a value, and optionally a datum hash.

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

isPubKeyOut :: TxOut -> Bool

Whether the output is a pay-to-pubkey output.

isPayToScriptOut :: TxOut -> Bool

Whether the output is a pay-to-script output.

outAddress :: Lens' TxOut Address

The address of a transaction output.

outValue :: Lens' TxOut Value

The value of a transaction output. | TODO: Compute address again

txOutPubKey :: TxOut -> Maybe PubKeyHash

The public key attached to a TxOut, if there is one.

txOutDatum :: TxOut -> Maybe DatumHash

The datum attached to a TxOut, if there is one.

pubKeyHashTxOut :: Value -> PubKeyHash -> TxOut

Create a transaction output locked by a public key.

Transaction inputs

data TxInType

The type of a transaction input.

Constructors

ConsumeScriptAddress !Validator !Redeemer !Datum
A transaction input that consumes a script address with the given validator, redeemer, and datum.

ConsumePublicKeyAddress
A transaction input that consumes a public key address.

ConsumeSimpleScriptAddress
Consume a simple script

data TxIn

A transaction input, consisting of a transaction output reference and an input type.

Constructors

TxIn

Fields

txInRef :: !TxOutRef

txInType :: Maybe TxInType

inRef :: Lens' TxIn TxOutRef

The TxOutRef spent by a transaction input.

inType :: Lens' TxIn (Maybe TxInType)

The type of a transaction input.

inScripts :: TxIn -> Maybe (Validator, Redeemer, Datum)

Validator, redeemer, and data scripts of a transaction input that spends a "pay to script" output.

pubKeyTxIn :: TxOutRef -> TxIn

A transaction input that spends a "pay to public key" output, given the witness.

scriptTxIn :: TxOutRef -> Validator -> Redeemer -> Datum -> TxIn

A transaction input that spends a "pay to script" output, given witnesses.

pubKeyTxIns :: Fold (Set TxIn) TxIn

Filter to get only the pubkey inputs.

scriptTxIns :: Fold (Set TxIn) TxIn

Filter to get only the script inputs.