Plutus.V2.Ledger.EvaluationContext

Documentation

type CostModelParams = Map Text Integer

data CostModelApplyError

The type of errors that applyParams can throw.

Constructors

CMUnknownParamError Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than recomputed on every evaluation.

There are two sets of parameters: one is with immediate unlifting and the other one is with deferred unlifting. We have to keep both of them, because depending on the language version either one has to be used or the other. We also compile them separately due to all the inlining and optimization that need to happen for things to be efficient.

toMachineParameters :: ProtocolVersion -> EvaluationContext -> DefaultMachineParameters

mkEvaluationContext :: MonadError CostModelApplyError m => CostModelParams -> m EvaluationContext

Build the EvaluationContext.

The input is a Map of strings to cost integer values (aka CostModelParams, CostModel) See Note [Inlining meanings of builtins].

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Comparably expensive to mkEvaluationContext, so it should only be used sparingly.

costModelParamsForTesting :: CostModelParams

The raw cost model params, only to be used for testing purposes.

evalCtxForTesting :: EvaluationContext

only to be for testing purposes: make an evaluation context by applying an empty set of protocol parameters

costModelParamNames :: Set Text

The set of valid names that a cost model parameter can take for this language version. It is used for the deserialization of CostModelParams.