PlutusCore.Executable.AstIO

Description

Reading and writing ASTs with various name types in flat format.

loadUplcASTfromSerialised :: Input -> IO (UplcProg ())

loadUplcASTfromHex :: Input -> IO (UplcProg ())

fromNamedDeBruijnUPLC :: UplcProgNDB ann -> UplcProg ann

Convert an untyped program with named de Bruijn indices to one with textual names.

toDeBruijnTermPLC :: PlcTerm ann -> PlcTermDB ann

Convert a typed term to one where the name type is de Bruijn indices.

toDeBruijnTermUPLC :: UplcTerm ann -> UplcTermDB ann

Convert an untyped term to one where the name type is de Bruijn indices.

toDeBruijnUPLC :: UplcProg ann -> UplcProgDB ann

Convert an untyped program to one where the name type is de Bruijn indices.

toDeBruijnTypePLC :: PlcType ann -> PlcTypeDB ann

Convert a type to one where the tyname type is de Bruijn indices.

toNamedDeBruijnUPLC :: UplcProg ann -> UplcProgNDB ann

Convert an untyped program to one where the name type is textual names with de Bruijn indices.