PlutusCore.Executable.Common

getUplcExamples :: IO [(ExampleName, SomeExample)]

Get available untyped examples. Currently the untyped examples are obtained by erasing typed ones, but it might be useful to have some untyped ones that can't be obtained by erasure.

helpText

Arguments

:: SrcSpan)

Parse a program. The first argument (normally the file path) describes the input stream, the second is the program text.

printBudgetState :: PrintBudgetState cost => PrettyConfigPlc (p ann)) => Output -> Format -> PrintMode -> p ann -> IO ()

writeToOutput :: Show a => Output -> a -> IO ()