PlutusLedgerApi.Common.Versions

Description

This module contains the code for handling the various kinds of version that we care about:

Protocol versions

Plutus ledger languages

Plutus Core language versions

Cardano Protocol versions

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion :: Int

Protocol Version aliases

Based on https://github.com/IntersectMBO/cardano-ledger/wiki/First-Block-of-Each-Era

shelleyPV :: MajorProtocolVersion

Shelley era was introduced in protocol version 2.0

allegraPV :: MajorProtocolVersion

Allegra era was introduced in protocol version 3.0

maryPV :: MajorProtocolVersion

Mary era was introduced in protocol version 4.0

alonzoPV :: MajorProtocolVersion

Alonzo era was introduced in protocol version 5.0

vasilPV :: MajorProtocolVersion

Vasil era was introduced in protocol version 7.0

valentinePV :: MajorProtocolVersion

Protocol version 8.0 was the Valentine intra-era HF

conwayPV :: MajorProtocolVersion

Conway era was introduced in protocol version 9.0

knownPVs :: Set MajorProtocolVersion

The set of protocol versions that are "known", i.e. that have been released and have actual differences associated with them.

futurePV :: MajorProtocolVersion

This is a placeholder for when we don't yet know what protocol version will be used for something. It's a very high protocol version that should never appear in reality.

We should not assign names to future protocol versions until it's confirmed that they are correct, otherwise we could accidentally associate something with the wrong protocol version.

Plutus ledger languages

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in shelley era

PlutusV2
introduced in vasil era

PlutusV3
not yet enabled

Plutus Core language versions

data Version

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

Version-testing functions

ledgerLanguageIntroducedIn :: PlutusLedgerLanguage -> MajorProtocolVersion

Query the protocol version that a specific Plutus ledger language was first introduced in.

ledgerLanguagesAvailableIn :: MajorProtocolVersion -> Set PlutusLedgerLanguage

Which Plutus language versions are available in the given MajorProtocolVersion?

See Note [New builtins/language versions and protocol versions]

plcVersionsIntroducedIn :: Map (PlutusLedgerLanguage, MajorProtocolVersion) (Set Version)

A map indicating which Plutus Core versions were introduced in which MajorProtocolVersion and PlutusLedgerLanguage. Each version should appear at most once.

This must be updated when new versions are added. See Note [New builtins/language versions and protocol versions]

plcVersionsAvailableIn :: PlutusLedgerLanguage -> MajorProtocolVersion -> Set Version

Which Plutus Core language versions are available in the given PlutusLedgerLanguage and MajorProtocolVersion?

See Note [New builtins/language versions and protocol versions]

builtinsIntroducedIn :: Map (PlutusLedgerLanguage, MajorProtocolVersion) (Set DefaultFun)

A map indicating which builtin functions were introduced in which MajorProtocolVersion. Each builtin function should appear at most once.

This must be updated when new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn :: PlutusLedgerLanguage -> MajorProtocolVersion -> Set DefaultFun

Which builtin functions are available in the given given PlutusLedgerLanguage and MajorProtocolVersion?

See Note [New builtins/language versions and protocol versions]