PlutusLedgerApi.Test.EvaluationEvent

Documentation

data ScriptEvaluationEvents

This type contains a list of on-chain script evaluation events. All PlutusV1 evaluations (if any) share the same cost parameters. Same with PlutusV2.

Sharing the cost parameters lets us avoid creating a new EvaluationContext for each ScriptEvaluationEvent.

Constructors

ScriptEvaluationEvents

Fields

eventsCostParamsV1 :: Decoder s [ScriptEvaluationEvents]

type Rep ScriptEvaluationEvents

Instance details

Defined in PlutusLedgerApi.Test.EvaluationEvent

type Rep ScriptEvaluationEvents = D1 ('MetaData "ScriptEvaluationEvents" "PlutusLedgerApi.Test.EvaluationEvent" "plutus-ledger-api-1.60.0.0-inplace-plutus-ledger-api-testlib" 'False) (C1 ('MetaCons "ScriptEvaluationEvents" 'PrefixI 'True) (S1 ('MetaSel ('Just "eventsCostParamsV1") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Maybe [Int64])) :*: (S1 ('MetaSel ('Just "eventsCostParamsV2") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Maybe [Int64])) :*: S1 ('MetaSel ('Just "eventsEvents") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (NonEmpty ScriptEvaluationEvent)))))

data ScriptEvaluationEvent

Information about an on-chain script evaluation event, specifically the information needed to evaluate the script, and the expected result.

Constructors

PlutusEvent PlutusLedgerLanguage) :*: (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ScriptEvaluationData) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ScriptEvaluationResult))))

data ScriptEvaluationData

All the data needed to evaluate a script using the ledger API, except for the cost model parameters, as these are tracked separately.

Constructors

ScriptEvaluationData

Fields

dataProtocolVersion :: EvaluationContext

-> [Int64]
Cost parameters

-> ScriptEvaluationEvent

-> Maybe UnexpectedEvaluationResult

Re-evaluate an on-chain script evaluation event.