PlutusLedgerApi.V1

Description

The interface to Plutus V1 for the ledger.

Scripts

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

data Script

A script on the chain. This is an opaque type as far as the chain is concerned.

fromCompiledCode :: CompiledCode a -> Script

Turn a CompiledCode (usually produced by compile) into a Script for use with this package.

Validating scripts

isScriptWellFormed :: ProtocolVersion -> SerialisedScript -> Bool

Check if a Script is "valid" according to a protocol version. At the moment this means "deserialises correctly", which in particular implies that it is (almost certainly) an encoded script and the script does not mention any builtins unavailable in the given protocol version.

Running scripts

evaluateScriptRestricting

Arguments

:: ProtocolVersion

-> VerboseMode
Whether to produce log output

-> EvaluationContext
The cost model that should already be synced to the most recent cost-model-params coming from the current protocol

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

evaluateScriptCounting

Arguments

:: ProtocolVersion

-> VerboseMode
Whether to produce log output

-> EvaluationContext
The cost model that should already be synced to the most recent cost-model-params coming from the current protocol

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Protocol version

data ProtocolVersion

This represents the Cardano protocol version, with its major and minor components. This relies on careful understanding between us and the ledger as to what this means.

Constructors

ProtocolVersion

Fields

pvMajor :: Int

pvMinor :: Int

Verbose mode and log output

data VerboseMode

A simple toggle indicating whether or not we should produce logs.

Constructors

Verbose

Quiet

type LogOutput = [Text]

The type of log output: just a list of Text.

Costing-related types

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

newtype ExCPU

Counts CPU units in picoseconds: maximum value for SatInt is 2^63 ps, or appproximately 106 days.

Constructors

ExCPU CostingInteger

newtype ExMemory

Counts size in machine words.

Constructors

ExMemory CostingInteger

data SatInt

Cost model

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than recomputed on every evaluation.

There are two sets of parameters: one is with immediate unlifting and the other one is with deferred unlifting. We have to keep both of them, because depending on the language version either one has to be used or the other. We also compile them separately due to all the inlining and optimization that need to happen for things to be efficient.

mkEvaluationContext :: MonadError CostModelApplyError m => [Integer] -> m EvaluationContext

Build the EvaluationContext.

The input is a list of integer values passed from the ledger and are expected to appear in correct order.

data ParamName

The enumeration of all possible cost model parameter names for this language version. IMPORTANT: The order of appearance of the data constructors here matters. DO NOT REORDER. See Note [Quotation marks in cost model parameter constructors]

Constructors

AddInteger'cpu'arguments'intercept

AddInteger'cpu'arguments'slope

AddInteger'memory'arguments'intercept

AddInteger'memory'arguments'slope

AppendByteString'cpu'arguments'intercept

AppendByteString'cpu'arguments'slope

AppendByteString'memory'arguments'intercept

AppendByteString'memory'arguments'slope

AppendString'cpu'arguments'intercept

AppendString'cpu'arguments'slope

AppendString'memory'arguments'intercept

AppendString'memory'arguments'slope

BData'cpu'arguments

BData'memory'arguments

Blake2b_256'cpu'arguments'intercept

Blake2b_256'cpu'arguments'slope

Blake2b_256'memory'arguments

CekApplyCost'exBudgetCPU

CekApplyCost'exBudgetMemory

CekBuiltinCost'exBudgetCPU

CekBuiltinCost'exBudgetMemory

CekConstCost'exBudgetCPU

CekConstCost'exBudgetMemory

CekDelayCost'exBudgetCPU

CekDelayCost'exBudgetMemory

CekForceCost'exBudgetCPU

CekForceCost'exBudgetMemory

CekLamCost'exBudgetCPU

CekLamCost'exBudgetMemory

CekStartupCost'exBudgetCPU

CekStartupCost'exBudgetMemory

CekVarCost'exBudgetCPU

CekVarCost'exBudgetMemory

ChooseData'cpu'arguments

ChooseData'memory'arguments

ChooseList'cpu'arguments

ChooseList'memory'arguments

ChooseUnit'cpu'arguments

ChooseUnit'memory'arguments

ConsByteString'cpu'arguments'intercept

ConsByteString'cpu'arguments'slope

ConsByteString'memory'arguments'intercept

ConsByteString'memory'arguments'slope

ConstrData'cpu'arguments

ConstrData'memory'arguments

DecodeUtf8'cpu'arguments'intercept

DecodeUtf8'cpu'arguments'slope

DecodeUtf8'memory'arguments'intercept

DecodeUtf8'memory'arguments'slope

DivideInteger'cpu'arguments'constant

DivideInteger'cpu'arguments'model'arguments'intercept

DivideInteger'cpu'arguments'model'arguments'slope

DivideInteger'memory'arguments'intercept

DivideInteger'memory'arguments'minimum

DivideInteger'memory'arguments'slope

EncodeUtf8'cpu'arguments'intercept

EncodeUtf8'cpu'arguments'slope

EncodeUtf8'memory'arguments'intercept

EncodeUtf8'memory'arguments'slope

EqualsByteString'cpu'arguments'constant

EqualsByteString'cpu'arguments'intercept

EqualsByteString'cpu'arguments'slope

EqualsByteString'memory'arguments

EqualsData'cpu'arguments'intercept

EqualsData'cpu'arguments'slope

EqualsData'memory'arguments

EqualsInteger'cpu'arguments'intercept

EqualsInteger'cpu'arguments'slope

EqualsInteger'memory'arguments

EqualsString'cpu'arguments'constant

EqualsString'cpu'arguments'intercept

EqualsString'cpu'arguments'slope

EqualsString'memory'arguments

FstPair'cpu'arguments

FstPair'memory'arguments

HeadList'cpu'arguments

HeadList'memory'arguments

IData'cpu'arguments

IData'memory'arguments

IfThenElse'cpu'arguments

IfThenElse'memory'arguments

IndexByteString'cpu'arguments

IndexByteString'memory'arguments

LengthOfByteString'cpu'arguments

LengthOfByteString'memory'arguments

LessThanByteString'cpu'arguments'intercept

LessThanByteString'cpu'arguments'slope

LessThanByteString'memory'arguments

LessThanEqualsByteString'cpu'arguments'intercept

LessThanEqualsByteString'cpu'arguments'slope

LessThanEqualsByteString'memory'arguments

LessThanEqualsInteger'cpu'arguments'intercept

LessThanEqualsInteger'cpu'arguments'slope

LessThanEqualsInteger'memory'arguments

LessThanInteger'cpu'arguments'intercept

LessThanInteger'cpu'arguments'slope

LessThanInteger'memory'arguments

ListData'cpu'arguments

ListData'memory'arguments

MapData'cpu'arguments

MapData'memory'arguments

MkCons'cpu'arguments

MkCons'memory'arguments

MkNilData'cpu'arguments

MkNilData'memory'arguments

MkNilPairData'cpu'arguments

MkNilPairData'memory'arguments

MkPairData'cpu'arguments

MkPairData'memory'arguments

ModInteger'cpu'arguments'constant

ModInteger'cpu'arguments'model'arguments'intercept

ModInteger'cpu'arguments'model'arguments'slope

ModInteger'memory'arguments'intercept

ModInteger'memory'arguments'minimum

ModInteger'memory'arguments'slope

MultiplyInteger'cpu'arguments'intercept

MultiplyInteger'cpu'arguments'slope

MultiplyInteger'memory'arguments'intercept

MultiplyInteger'memory'arguments'slope

NullList'cpu'arguments

NullList'memory'arguments

QuotientInteger'cpu'arguments'constant

QuotientInteger'cpu'arguments'model'arguments'intercept

QuotientInteger'cpu'arguments'model'arguments'slope

QuotientInteger'memory'arguments'intercept

QuotientInteger'memory'arguments'minimum

QuotientInteger'memory'arguments'slope

RemainderInteger'cpu'arguments'constant

RemainderInteger'cpu'arguments'model'arguments'intercept

RemainderInteger'cpu'arguments'model'arguments'slope

RemainderInteger'memory'arguments'intercept

RemainderInteger'memory'arguments'minimum

RemainderInteger'memory'arguments'slope

Sha2_256'cpu'arguments'intercept

Sha2_256'cpu'arguments'slope

Sha2_256'memory'arguments

Sha3_256'cpu'arguments'intercept

Sha3_256'cpu'arguments'slope

Sha3_256'memory'arguments

SliceByteString'cpu'arguments'intercept

SliceByteString'cpu'arguments'slope

SliceByteString'memory'arguments'intercept

SliceByteString'memory'arguments'slope

SndPair'cpu'arguments

SndPair'memory'arguments

SubtractInteger'cpu'arguments'intercept

SubtractInteger'cpu'arguments'slope

SubtractInteger'memory'arguments'intercept

SubtractInteger'memory'arguments'slope

TailList'cpu'arguments

TailList'memory'arguments

Trace'cpu'arguments

Trace'memory'arguments

UnBData'cpu'arguments

UnBData'memory'arguments

UnConstrData'cpu'arguments

UnConstrData'memory'arguments

UnIData'cpu'arguments

UnIData'memory'arguments

UnListData'cpu'arguments

UnListData'memory'arguments

UnMapData'cpu'arguments

UnMapData'memory'arguments

VerifyEd25519Signature'cpu'arguments'intercept

VerifyEd25519Signature'cpu'arguments'slope

VerifyEd25519Signature'memory'arguments

data CostModelApplyError

The type of errors that applyParams can throw.

Constructors

CMUnknownParamError Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

CMWrongNumberOfParams Int Int
the ledger is supposed to pass the full list of params, no more, no less params.

type CostModelParams = Map Text Integer

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Comparably expensive to mkEvaluationContext, so it should only be used sparingly.

Context types

data ScriptContext

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

scriptContextPurpose :: ScriptPurpose

data ScriptPurpose

Purpose of the script that is currently running

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

Supporting types used in the context types

ByteStrings

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

toBuiltin :: ToBuiltin a arep => a -> arep

fromBuiltin :: FromBuiltin arep a => arep -> a

Bytes

newtype LedgerBytes

Constructors

LedgerBytes

Fields

getLedgerBytes :: BuiltinByteString

fromBytes :: ByteString -> LedgerBytes

Certificates

data DCert

A representation of the ledger DCert. Some information is digested, and not included

Constructors

DCertDelegRegKey StakingCredential

DCertDelegDeRegKey StakingCredential

DCertDelegDelegate

Fields

StakingCredential

delegator

PubKeyHash

delegatee

DCertPoolRegister
A digest of the PoolParams

Fields

PubKeyHash

poolId

PubKeyHash

pool VFR

DCertPoolRetire PubKeyHash Integer
The retiremant certificate and the Epoch N

DCertGenesis
A really terse Digest

DCertMir
Another really terse Digest

Credentials

data StakingCredential

Staking credential used to assign rewards

Constructors

StakingHash Credential

StakingPtr Integer Integer Integer

data Credential

Credential required to unlock a transaction output

Constructors

PubKeyCredential PubKeyHash
The transaction that spends this output must be signed by the private key

ScriptCredential ValidatorHash
The transaction that spends this output must include the validator script and be accepted by the validator.

Value

newtype Value

A cryptocurrency value. This is a map from CurrencySymbols to a quantity of that currency.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

See note [Currencies] for more details.

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Integer)

newtype CurrencySymbol

Constructors

CurrencySymbol

Fields

unCurrencySymbol :: BuiltinByteString

newtype TokenName

ByteString of a name of a token, shown as UTF-8 string when possible

Constructors

TokenName

Fields

unTokenName :: BuiltinByteString

singleton :: CurrencySymbol -> TokenName -> Integer -> Value

Make a Value containing only the given quantity of the given currency.

unionWith :: (Integer -> Integer -> Integer) -> Value -> Value -> Value

adaSymbol :: CurrencySymbol

The CurrencySymbol of the Ada currency.

adaToken :: TokenName

The TokenName of the Ada currency.

Time

newtype POSIXTime

POSIX time is measured as the number of milliseconds since 1970-01-01T00:00:00Z

Constructors

POSIXTime

Fields

getPOSIXTime :: Integer

type POSIXTimeRange = Interval POSIXTime

An Interval of POSIXTimes.

Types for representing transactions

data Address

Address with two kinds of credentials, normal and staking.

Constructors

Address

Fields

addressCredential :: Credential

addressStakingCredential :: Maybe StakingCredential

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself.

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

newtype TxId

A transaction ID, using a SHA256 hash as the transaction id.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxInfo

A pending transaction. This is the view as seen by validator scripts, so some details are stripped out.

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

Transaction inputs

txInfoOutputs :: [TxOut]

Transaction outputs

txInfoFee :: Value

The fee paid by this transaction.

txInfoMint :: Value

The Value minted by this transaction.

txInfoDCert :: [DCert]

Digests of certificates included in this transaction

txInfoWdrl :: [(StakingCredential, Integer)]

Withdrawals

txInfoValidRange :: POSIXTimeRange

The valid range for the transaction.

txInfoSignatories :: [PubKeyHash]

Signatures provided with the transaction, attested that they all signed the tx

txInfoData :: [(DatumHash, Datum)]

txInfoId :: TxId

Hash of the pending transaction (excluding witnesses)

data TxOut

A transaction output, consisting of a target address, a value, and optionally a datum hash.

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

Intervals

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data Extended a

A set extended with a positive and negative infinity.

Constructors

NegInf

Finite a

PosInf

type Closure = Bool

Whether a bound is inclusive or not.

data UpperBound a

The upper bound of an interval.

Constructors

UpperBound (Extended a) Closure

data LowerBound a

The lower bound of an interval.

Constructors

LowerBound (Extended a) Closure

always :: Interval a

An Interval that covers every slot.

from :: a -> Interval a

from a is an Interval that includes all values that are greater than or equal to a.

to :: a -> Interval a

to a is an Interval that includes all values that are smaller than or equal to a.

lowerBound :: a -> LowerBound a

upperBound :: a -> UpperBound a

strictLowerBound :: a -> LowerBound a

strictUpperBound :: a -> UpperBound a

Newtypes for script/datum types and hash types

newtype Validator

Validator is a wrapper around Scripts which are used as validators in transaction outputs.

Constructors

Validator

Fields

getValidator :: Script

mkValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> ()) -> Validator

unValidatorScript :: Validator -> Script

newtype ValidatorHash

Script runtime representation of a Digest SHA256.

Constructors

ValidatorHash BuiltinByteString

newtype MintingPolicy

MintingPolicy is a wrapper around Scripts which are used as validators for minting constraints.

Constructors

MintingPolicy

Fields

getMintingPolicy :: Script

mkMintingPolicyScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> MintingPolicy

unMintingPolicyScript :: MintingPolicy -> Script

newtype MintingPolicyHash

Script runtime representation of a Digest SHA256.

Constructors

MintingPolicyHash BuiltinByteString

newtype StakeValidator

StakeValidator is a wrapper around Scripts which are used as validators for withdrawals and stake address certificates.

Constructors

StakeValidator

Fields

getStakeValidator :: Script

mkStakeValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> StakeValidator

unStakeValidatorScript :: StakeValidator -> Script

newtype StakeValidatorHash

Script runtime representation of a Digest SHA256.

Constructors

StakeValidatorHash BuiltinByteString

newtype Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

Constructors

Redeemer

Fields

getRedeemer :: BuiltinData

newtype RedeemerHash

Script runtime representation of a Digest SHA256.

Constructors

RedeemerHash BuiltinByteString

newtype Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

Constructors

Datum

Fields

getDatum :: BuiltinData

newtype DatumHash

Script runtime representation of a Digest SHA256.

Constructors

DatumHash BuiltinByteString

Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData Data

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

class FromData a where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData a where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

toData :: ToData a => a -> Data

Convert a value to Data.

fromData :: FromData a => Data -> Maybe a

Convert a value from Data, returning Nothing if this fails.

dataToBuiltinData :: Data -> BuiltinData

Convert a Data into a BuiltinData. Only works off-chain.

builtinDataToData :: BuiltinData -> Data

Convert a BuiltinData into a Data. Only works off-chain.

Errors

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError DeserialiseFailure
A serialisation error

IncompatibleVersionError (Version ())
An error indicating a version tag that we don't support TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected