PlutusLedgerApi.V1.Contexts

Pending transactions and related types

data TxInfo

A pending transaction. This is the view as seen by validator scripts, so some details are stripped out.

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

Transaction inputs; cannot be an empty list

txInfoOutputs :: [TxOut]

Transaction outputs

txInfoFee :: Value

The fee paid by this transaction.

txInfoMint :: Value

The Value minted by this transaction.

txInfoDCert :: [DCert]

Digests of certificates included in this transaction

txInfoWdrl :: [(StakingCredential, Integer)]

Withdrawals

txInfoValidRange :: POSIXTimeRange

The valid range for the transaction.

txInfoSignatories :: [PubKeyHash]

Signatures provided with the transaction, attested that they all signed the tx

txInfoData :: [(DatumHash, Datum)]

The lookup table of datums attached to the transaction

txInfoId :: TxId

Hash of the pending transaction body (i.e. transaction excluding witnesses)

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextPurpose :: ScriptPurpose

the purpose of the currently-executing script

data ScriptPurpose

Purpose of the script that is currently running

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxOut

A transaction output, consisting of a target address (Address), a value (Value), and optionally a datum hash (DatumHash).

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

The transaction ID.

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

findOwnInput :: ScriptContext -> Maybe TxInInfo

Find the input currently being validated.

findDatum :: DatumHash -> TxInfo -> Maybe Datum

Find the data corresponding to a data hash, if there is one

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

Find the hash of a datum, if it is part of the pending transaction's hashes

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

Given a UTXO reference and a transaction (TxInfo), resolve it to one of the transaction's inputs (TxInInfo).

findContinuingOutputs :: ScriptContext -> [Integer]

Finds all the outputs that pay to the same script address that we are currently spending from, if any.

getContinuingOutputs :: ScriptContext -> [TxOut]

Get all the outputs that pay to the same script address we are currently spending from, if any.

Validator functions

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> [Value]

Get the values paid to a public key address by a pending transaction.

valuePaidTo :: TxInfo -> PubKeyHash -> Value

Get the total value paid to a public key address by a pending transaction.

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)

txSignedBy :: TxInfo -> PubKeyHash -> Bool

Check if a transaction was signed by the given public key.

valueSpent :: TxInfo -> Value

Get the total value of inputs spent by this transaction.

valueProduced :: TxInfo -> Value

Get the total value of outputs produced by this transaction.

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

The CurrencySymbol of the current validator script.