PlutusLedgerApi.V1.Credential

Description

Address and staking address credentials for outputs.

Documentation

data StakingCredential

Staking credential used to assign rewards.

Constructors

StakingHash Credential
The staking hash is the Credential required to unlock a transaction output. Either a public key credential (PubKeyHash) or a script credential (ScriptHash). Both are hashed with BLAKE2b-244. 28 byte.

StakingPtr
The certificate pointer, constructed by the given slot number, transaction and certificate indices. NB: The fields should really be all Word64, as they are implemented in Word64, but RepIsStuckError StakingCredential :: Type -> Type) (Rep StakingCredential)))

data Credential

Credentials required to unlock a transaction output.

Constructors

PubKeyCredential PubKeyHash
The transaction that spends this output must be signed by the private key. See PubKeyHash.

ScriptCredential ScriptHash
The transaction that spends this output must include the validator script and be accepted by the validator. See ScriptHash.