PlutusLedgerApi.V1.Data.Address

Documentation

data Address

An address may contain two credentials, the payment credential and optionally a StakingCredential.

pattern Address :: Credential -> Maybe StakingCredential -> Address

matchAddress :: forall r. Address -> (Credential -> Maybe StakingCredential -> r) -> r

addressCredential :: Address -> Credential

addressStakingCredential :: Address -> Maybe StakingCredential

pubKeyHashAddress :: PubKeyHash -> Address

The address that should be targeted by a transaction output locked by the public key with the given hash.

toPubKeyHash :: Address -> Maybe PubKeyHash

The PubKeyHash of the address, if any

toScriptHash :: Address -> Maybe ScriptHash

The validator hash of the address, if any

scriptHashAddress :: ScriptHash -> Address

The address that should be used by a transaction output locked by the given validator script hash.

stakingCredential :: Address -> Maybe StakingCredential

The staking credential of an address (if any)