PlutusLedgerApi.V1.Interval

Description

A type for intervals and associated functions.

Documentation

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

The Equality of two intervals is specified as the canonical, structural equality and not the equality of the elements of their two underlying sets; the same holds for Ord.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data UpperBound a

The upper bound of an interval.

Constructors

UpperBound (Extended a) Closure

data LowerBound a

The lower bound of an interval.

Constructors

LowerBound (Extended a) Closure

data Extended a

A set extended with a positive and negative infinity.

Constructors

NegInf

Finite a

PosInf

type Closure = Bool

Whether a bound is inclusive or not.

member :: Ord a => a -> Interval a -> Bool

Check whether a value is in an interval.

interval :: a -> a -> Interval a

interval a b includes all values that are greater than or equal to a and smaller than or equal to b. Therefore it includes a and b. In math. notation: [a,b]

from :: a -> Interval a

from a is an Interval that includes all values that are greater than or equal to a. In math. notation: [a,+∞]

to :: a -> Interval a

to a is an Interval that includes all values that are smaller than or equal to a. In math. notation: [-∞,a]

always :: Interval a

An Interval that covers every slot. In math. notation [-∞,+∞]

never :: Interval a

An Interval that is empty.

There can be many empty intervals, see isEmpty. The empty interval never is arbitrarily set to [+∞,-∞].

singleton :: a -> Interval a

Create an interval that includes just a single concrete point a, i.e. having the same non-strict lower and upper bounds. In math.notation: [a,a]

hull :: Ord a => Interval a -> Interval a -> Interval a

'hull a b' is the smallest interval containing a and b.

intersection :: Ord a => Interval a -> Interval a -> Interval a

'intersection a b' is the largest interval that is contained in a and in b, if it exists.

overlaps :: (Enum a, Ord a) => Interval a -> Interval a -> Bool

Check whether two intervals overlap, that is, whether there is a value that is a member of both intervals.

contains :: Ord a => Interval a -> Interval a -> Bool

a contains b is true if the Interval b is entirely contained in a. That is, a contains b if for every entry s, if member s b then member s a.

isEmpty :: (Enum a, Ord a) => Interval a -> Bool

Check if an Interval is empty.

There can be many empty intervals, given a set of values. Two Intervals being empty does not imply that they are Equal to each other.

before :: Ord a => a -> Interval a -> Bool

Check if a value is earlier than the beginning of an Interval.

after :: Ord a => a -> Interval a -> Bool

Check if a value is later than the end of an Interval.

lowerBound :: a -> LowerBound a

Construct a lower bound from a value.

The resulting bound includes all values that are equal or greater than the input value.

upperBound :: a -> UpperBound a

Construct an upper bound from a value.

The resulting bound includes all values that are equal or smaller than the input value.

strictLowerBound :: a -> LowerBound a

Construct a strict lower bound from a value.

The resulting bound includes all values that are (strictly) greater than the input value.

strictUpperBound :: a -> UpperBound a

Construct a strict upper bound from a value.

The resulting bound includes all values that are (strictly) smaller than the input value.