PlutusLedgerApi.V1.Time

Description

UTCTime and UTCTime ranges.

Documentation

newtype POSIXTime

POSIX time is measured as the number of milliseconds since 1970-01-01T00:00:00Z. This is not the same as Haskell's POSIXTime

Constructors

POSIXTime

Fields

getPOSIXTime :: Integer

type POSIXTimeRange = Interval POSIXTime

An Interval of POSIXTimes.

newtype DiffMilliSeconds

This is a length of time, as measured by a number of milliseconds.

Constructors

DiffMilliSeconds Integer

fromMilliSeconds :: DiffMilliSeconds -> POSIXTime

Simple conversion from DiffMilliSeconds to POSIXTime.