PlutusLedgerApi.V1.Value

Description

Functions for working with Value.

Currency symbols

newtype CurrencySymbol

ByteString representing the currency, hashed with BLAKE2b-224. It is empty for Ada, 28 bytes for MintingPolicyHash. Forms an AssetClass along with TokenName. A Value is a map from CurrencySymbol's to a map from TokenName to an Integer.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

CurrencySymbol

Fields

unCurrencySymbol :: BuiltinByteString

currencySymbol :: ByteString -> CurrencySymbol

Creates CurrencySymbol from raw ByteString.

adaSymbol :: CurrencySymbol

The CurrencySymbol of the Ada currency.

Token names

newtype TokenName

ByteString of a name of a token. Shown as UTF-8 string when possible. Should be no longer than 32 bytes, empty for Ada. Forms an AssetClass along with a CurrencySymbol.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TokenName

Fields

unTokenName :: BuiltinByteString

tokenName :: ByteString -> TokenName

Creates TokenName from raw ByteString.

toString :: TokenName -> String

Turn a TokenName to a hex-encoded String

Compared to show , it will not surround the string with double-quotes.

adaToken :: TokenName

The TokenName of the Ada currency.

Asset classes

newtype AssetClass

An asset class, identified by a CurrencySymbol and a TokenName.

Constructors

AssetClass

Fields

unAssetClass :: (CurrencySymbol, TokenName)

assetClass :: CurrencySymbol -> TokenName -> AssetClass

The curried version of AssetClass constructor

assetClassValue :: AssetClass -> Integer -> Value

A Value containing the given amount of the asset class.

assetClassValueOf :: Value -> AssetClass -> Integer

Get the quantity of the given AssetClass class in the Value.

Value

newtype Value

The Value type represents a collection of amounts of different currencies. We can think of Value as a vector space whose dimensions are currencies.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

There is no 'Ord Value' instance since Value is only a partial order, so compare can't do the right thing in some cases.

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Integer)

singleton :: CurrencySymbol -> TokenName -> Integer -> Value

Make a Value containing only the given quantity of the given currency.

valueOf :: Value -> CurrencySymbol -> TokenName -> Integer

Get the quantity of the given currency in the Value. Assumes that the underlying map doesn't contain duplicate keys.

currencySymbolValueOf :: Value -> CurrencySymbol -> Integer

Get the total value of the currency symbol in the Value map. Assumes that the underlying map doesn't contain duplicate keys.

lovelaceValue :: Lovelace -> Value

A Value containing the given quantity of Lovelace.

lovelaceValueOf :: Value -> Lovelace

Get the quantity of Lovelace in the Value.

scale :: Module s v => s -> v -> v

symbols :: Value -> [CurrencySymbol]

The list of CurrencySymbols of a Value.

Partial order operations

geq :: Value -> Value -> Bool

Check whether one Value is greater than or equal to another. See Value for an explanation of how operations on Values work.

gt :: Value -> Value -> Bool

Check whether one Value is strictly greater than another. This is *not* a pointwise operation. gt l r means geq l r && not (eq l r).

leq :: Value -> Value -> Bool

Check whether one Value is less than or equal to another. See Value for an explanation of how operations on Values work.

lt :: Value -> Value -> Bool

Check whether one Value is strictly less than another. This is *not* a pointwise operation. lt l r means leq l r && not (eq l r).

Etc.

isZero :: Value -> Bool

Check whether a Value is zero.

split :: Value -> (Value, Value)

Split a Value into its positive and negative parts. The first element of the tuple contains the negative parts of the Value, the second element contains the positive parts.

negate (fst (split a)) plus (snd (split a)) == a

unionWith :: (Integer -> Integer -> Integer) -> Value -> Value -> Value

Combine two Value maps with the argument function. Assumes the well-definedness of the two maps.

flattenValue :: Value -> [(CurrencySymbol, TokenName, Integer)]

Convert a Value to a simple list, keeping only the non-zero amounts. Note that the result isn't sorted, meaning v1 == v2 doesn't generally imply flattenValue v1 == flattenValue v2. Also assumes that there are no duplicate keys in the Value Map.

newtype Lovelace

Constructors

Lovelace

Fields

getLovelace :: Integer