PlutusLedgerApi.V2.Data.Tx

Transactions

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification. -- editorconfig-checker-disable-file

Constructors

TxId

Fields

getTxId :: BuiltinByteString

type Rep TxId

Instance details

Defined in PlutusLedgerApi.V1.Data.Tx

type Rep TxId = D1 ('MetaData "TxId" "PlutusLedgerApi.V1.Data.Tx" "plutus-ledger-api-1.60.0.0-inplace" 'True) (C1 ('MetaCons "TxId" 'PrefixI 'True) (S1 ('MetaSel ('Just "getTxId") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 BuiltinByteString)))

data ScriptTag

A tag indicating the type of script that we are pointing to.

See also ScriptPurpose

Constructors

Spend

Mint

Cert

Reward

data RedeemerPtr

A redeemer pointer is a pair of a script type tag (ScriptTag) t and an index i, picking out the i-th script of type t in the transaction.

Constructors

RedeemerPtr ScriptTag Integer

type Redeemers = Map RedeemerPtr Redeemer

Redeemers is a Map of redeemer pointer (RedeemerPtr) and its Redeemer.

Transaction outputs

data TxOut

A transaction output, consisting of a target address, a value, optionally a datum/datum hash, and optionally a reference script.

type Rep TxOut

Instance details

Defined in PlutusLedgerApi.V2.Data.Tx

type Rep TxOut = D1 ('MetaData "TxOut" "PlutusLedgerApi.V2.Data.Tx" "plutus-ledger-api-1.60.0.0-inplace" 'True) (C1 ('MetaCons "TxOut_6989586621680146156" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 BuiltinData)))

pattern TxOut :: Address -> Value -> OutputDatum -> Maybe ScriptHash -> TxOut

matchTxOut :: forall r. TxOut -> (Address -> Value -> OutputDatum -> Maybe ScriptHash -> r) -> r

txOutAddress :: TxOut -> Address

txOutValue :: TxOut -> Value

txOutDatum :: TxOut -> OutputDatum

txOutReferenceScript :: TxOut -> Maybe ScriptHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

type Rep TxOutRef

Instance details

Defined in PlutusLedgerApi.V1.Data.Tx

type Rep TxOutRef = D1 ('MetaData "TxOutRef" "PlutusLedgerApi.V1.Data.Tx" "plutus-ledger-api-1.60.0.0-inplace" 'True) (C1 ('MetaCons "TxOutRef_6989586621680120830" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 BuiltinData)))

pattern TxOutRef :: TxId -> Integer -> TxOutRef

txOutRefId :: TxOutRef -> TxId

txOutRefIdx :: TxOutRef -> Integer

data OutputDatum

The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").

txOutPubKey :: TxOut -> Lens' TxOut (Maybe ScriptHash)

The reference script attached to a TxOut.

pubKeyHashTxOut :: Value -> PubKeyHash -> TxOut

Create a transaction output locked by a public key.