PlutusLedgerApi.V2.EvaluationContext

Documentation

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than recomputed on every evaluation.

There are two sets of parameters: one is with immediate unlifting and the other one is with deferred unlifting. We have to keep both of them, because depending on the language version either one has to be used or the other. We also compile them separately due to all the inlining and optimization that need to happen for things to be efficient.

mkEvaluationContext :: MonadError CostModelApplyError m => [Integer] -> m EvaluationContext

Build the EvaluationContext.

The input is a list of integer values passed from the ledger and are expected to appear in correct order.

type CostModelParams = Map Text Integer

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Comparably expensive to mkEvaluationContext, so it should only be used sparingly.

machineParametersImmediate :: EvaluationContext -> DefaultMachineParameters

machineParametersDeferred :: EvaluationContext -> DefaultMachineParameters

toMachineParameters :: ProtocolVersion -> EvaluationContext -> DefaultMachineParameters

data CostModelApplyError

The type of errors that applyParams can throw.

Constructors

CMUnknownParamError Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

CMWrongNumberOfParams Int Int
the ledger is supposed to pass the full list of params, no more, no less params.