PlutusLedgerApi.V3

Description

The interface to Plutus V3 for the ledger.

serialiseUPLC :: DefaultFun () -> SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseScript

Arguments

:: forall m. MonadError ScriptDecodeError m

=> MajorProtocolVersion
which major protocol version the script was submitted in.

-> SerialisedScript
the script to deserialise.

-> m ScriptForEvaluation

The deserialization from a serialised script into a ScriptForEvaluation, ready to be evaluated on-chain. Called inside phase-1 validation (i.e., deserialisation error is a phase-1 error).

uncheckedDeserialiseUPLC :: SerialisedScript -> DefaultFun ()

Deserialises a SerialisedScript back into an AST. Does *not* do ledger-language-version-specific checks like for allowable builtins.

Running scripts

evaluateScriptRestricting

Arguments

:: MajorProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> ScriptForEvaluation
The script to evaluate

-> Data
The ScriptContext argument to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

evaluateScriptCounting

Arguments

:: MajorProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ScriptForEvaluation
The script to evaluate

-> Data
The ScriptContext argument to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

CIP-1694

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

data TxCert

Constructors

TxCertRegStaking Credential (Maybe Lovelace)
Register staking credential with an optional deposit amount

TxCertUnRegStaking Credential (Maybe Lovelace)
Un-Register staking credential with an optional refund amount

TxCertDelegStaking Credential Delegatee
Delegate staking credential to a Delegatee

TxCertRegDeleg Credential Delegatee Lovelace
Register and delegate staking credential to a Delegatee in one certificate. Note that deposit is mandatory.

TxCertRegDRep DRepCredential Lovelace
Register a DRep with a deposit value. The optional anchor is omitted.

TxCertUpdateDRep DRepCredential
Update a DRep. The optional anchor is omitted.

TxCertUnRegDRep DRepCredential Lovelace
UnRegister a DRep with mandatory refund value

TxCertPoolRegister
A digest of the PoolParams

Fields

PubKeyHash

poolId

PubKeyHash

pool VFR

TxCertPoolRetire PubKeyHash RepIsStuckError GovernanceAction :: Type -> Type) (Rep GovernanceAction)))

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl.

Invariant: This map is non-empty, and the keys are stored in ascending order.

This Data object has the following format (in pseudocode):

ChangedParametersData = Map ChangedIdData ChangedManyValueData ChangedIdData = I Integer ChangedManyValueData = ChangedSingleValueData | List[ChangedSingleValueData...] -- ^ an arbitrary-length, heterogeneous (integer or ratio) list of values (to support sub-parameters)

ChangedSingleValueData = I Integer -- a proposed integer value | List[I Integer, I Integer] -- a proposed numerator,denominator (ratio value) -- ^ a 2-exact element list; *BE CAREFUL* because this can be alternatively (ambiguously) interpreted -- as a many-value data (sub-parameter) of two integer single-value data.

, where Map,I,List are the constructors of Data and Integer is the usual arbitrary-precision PlutusTx/Haskell Integer.

Constructors

ChangedParameters

Fields

getChangedParameters :: BuiltinData

Protocol version

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion :: PlutusCore.Evaluation.Machine.ExBudget
type Rep ExBudget = D1 ('MetaData "ExBudget" "PlutusCore.Evaluation.Machine.ExBudget" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "ExBudget" 'PrefixI 'True) (S1 ('MetaSel ('Just "exBudgetCPU") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedUnpack) (Rec0 ExCPU) :*: S1 ('MetaSel ('Just "exBudgetMemory") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedUnpack) (Rec0 ExMemory)))

newtype ExCPU

Counts CPU units in picoseconds: maximum value for SatInt is 2^63 ps, or appproximately 106 days.

Constructors

ExCPU Data.SatInt

Associated Types

type Rep SatInt :: Type -> Type

Methods

from :: SatInt -> Rep SatInt x

to :: Rep SatInt x -> SatInt

Num SatInt
In the Num instance, we plug in our own addition, multiplication and subtraction function that perform overflow-checking.

Instance details

Defined in Data.SatInt

type Rep SatInt = D1 ('MetaData "SatInt" "Data.SatInt" "plutus-core-1.60.0.0-inplace-satint" 'True) (C1 ('MetaCons "SI" 'PrefixI 'True) (S1 ('MetaSel ('Just "unSatInt") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Int)))

fromSatInt :: Num a => SatInt -> a

An optimized version of fromIntegral . unSatInt.

Cost model

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

Different protocol versions may require different bundles of machine parameters, which allows us for example to tweak the shape of the costing function of a builtin, so that the builtin costs less. Currently this means that we have to create multiple DefaultMachineParameters per language version, which we put into a cache (represented by an association list) in order to avoid costly recomputation of machine parameters.

In order to get the appropriate DefaultMachineParameters at validation time we look it up in the cache using a semantics variant as a key. We compute the semantics variant from the protocol version using the stored function. Note that the semantics variant depends on the language version too, but the latter is known statically (because each language version has its own evaluation context), hence there's no reason to require it to be provided at runtime.

To say it differently, there's a matrix of semantics variants indexed by (LL, PV) pairs and we cache its particular row corresponding to the statically given LL in an EvaluationContext.

The reason why we associate a DefaultMachineParameters with a semantics variant rather than a protocol version are

generally there are far more protocol versions than semantics variants supported by a specific language version, so we save on pointless duplication of bundles of machine parameters

builtins don't know anything about protocol versions, only semantics variants. It is therefore more semantically precise to associate bundles of machine parameters with semantics variants than with protocol versions

mkEvaluationContext

Arguments

:: (MonadError CostModelApplyError m, MonadWriter [CostModelApplyWarn] m)

=> [Int64]
the (updated) cost model parameters of the protocol

-> m EvaluationContext

Build the EvaluationContext.

The input is a list of cost model parameters (which are integer values) passed from the ledger.

IMPORTANT: the cost model parameters MUST appear in the correct order, matching the names in ParamName. If the parameters are supplied in the wrong order then script cost calculations will be incorrect.

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

data ParamName

The enumeration of all possible cost model parameter names for this language version.

IMPORTANT: The order of appearance of the data constructors here matters. DO NOT REORDER. See Note [Quotation marks in cost model parameter constructors] See Note [Cost model parameters from the ledger's point of view]

Constructors

AddInteger'cpu'arguments'intercept

AddInteger'cpu'arguments'slope

AddInteger'memory'arguments'intercept

AddInteger'memory'arguments'slope

AppendByteString'cpu'arguments'intercept

AppendByteString'cpu'arguments'slope

AppendByteString'memory'arguments'intercept

AppendByteString'memory'arguments'slope

AppendString'cpu'arguments'intercept

AppendString'cpu'arguments'slope

AppendString'memory'arguments'intercept

AppendString'memory'arguments'slope

BData'cpu'arguments

BData'memory'arguments

Blake2b_256'cpu'arguments'intercept

Blake2b_256'cpu'arguments'slope

Blake2b_256'memory'arguments

CekApplyCost'exBudgetCPU

CekApplyCost'exBudgetMemory

CekBuiltinCost'exBudgetCPU

CekBuiltinCost'exBudgetMemory

CekConstCost'exBudgetCPU

CekConstCost'exBudgetMemory

CekDelayCost'exBudgetCPU

CekDelayCost'exBudgetMemory

CekForceCost'exBudgetCPU

CekForceCost'exBudgetMemory

CekLamCost'exBudgetCPU

CekLamCost'exBudgetMemory

CekStartupCost'exBudgetCPU

CekStartupCost'exBudgetMemory

CekVarCost'exBudgetCPU

CekVarCost'exBudgetMemory

ChooseData'cpu'arguments

ChooseData'memory'arguments

ChooseList'cpu'arguments

ChooseList'memory'arguments

ChooseUnit'cpu'arguments

ChooseUnit'memory'arguments

ConsByteString'cpu'arguments'intercept

ConsByteString'cpu'arguments'slope

ConsByteString'memory'arguments'intercept

ConsByteString'memory'arguments'slope

ConstrData'cpu'arguments

ConstrData'memory'arguments

DecodeUtf8'cpu'arguments'intercept

DecodeUtf8'cpu'arguments'slope

DecodeUtf8'memory'arguments'intercept

DecodeUtf8'memory'arguments'slope

DivideInteger'cpu'arguments'constant

DivideInteger'cpu'arguments'model'arguments'c00

DivideInteger'cpu'arguments'model'arguments'c01

DivideInteger'cpu'arguments'model'arguments'c02

DivideInteger'cpu'arguments'model'arguments'c10

DivideInteger'cpu'arguments'model'arguments'c11

DivideInteger'cpu'arguments'model'arguments'c20

DivideInteger'cpu'arguments'model'arguments'minimum

DivideInteger'memory'arguments'intercept

DivideInteger'memory'arguments'minimum

DivideInteger'memory'arguments'slope

EncodeUtf8'cpu'arguments'intercept

EncodeUtf8'cpu'arguments'slope

EncodeUtf8'memory'arguments'intercept

EncodeUtf8'memory'arguments'slope

EqualsByteString'cpu'arguments'constant

EqualsByteString'cpu'arguments'intercept

EqualsByteString'cpu'arguments'slope

EqualsByteString'memory'arguments

EqualsData'cpu'arguments'intercept

EqualsData'cpu'arguments'slope

EqualsData'memory'arguments

EqualsInteger'cpu'arguments'intercept

EqualsInteger'cpu'arguments'slope

EqualsInteger'memory'arguments

EqualsString'cpu'arguments'constant

EqualsString'cpu'arguments'intercept

EqualsString'cpu'arguments'slope

EqualsString'memory'arguments

FstPair'cpu'arguments

FstPair'memory'arguments

HeadList'cpu'arguments

HeadList'memory'arguments

IData'cpu'arguments

IData'memory'arguments

IfThenElse'cpu'arguments

IfThenElse'memory'arguments

IndexByteString'cpu'arguments

IndexByteString'memory'arguments

LengthOfByteString'cpu'arguments

LengthOfByteString'memory'arguments

LessThanByteString'cpu'arguments'intercept

LessThanByteString'cpu'arguments'slope

LessThanByteString'memory'arguments

LessThanEqualsByteString'cpu'arguments'intercept

LessThanEqualsByteString'cpu'arguments'slope

LessThanEqualsByteString'memory'arguments

LessThanEqualsInteger'cpu'arguments'intercept

LessThanEqualsInteger'cpu'arguments'slope

LessThanEqualsInteger'memory'arguments

LessThanInteger'cpu'arguments'intercept

LessThanInteger'cpu'arguments'slope

LessThanInteger'memory'arguments

ListData'cpu'arguments

ListData'memory'arguments

MapData'cpu'arguments

MapData'memory'arguments

MkCons'cpu'arguments

MkCons'memory'arguments

MkNilData'cpu'arguments

MkNilData'memory'arguments

MkNilPairData'cpu'arguments

MkNilPairData'memory'arguments

MkPairData'cpu'arguments

MkPairData'memory'arguments

ModInteger'cpu'arguments'constant

ModInteger'cpu'arguments'model'arguments'c00

ModInteger'cpu'arguments'model'arguments'c01

ModInteger'cpu'arguments'model'arguments'c02

ModInteger'cpu'arguments'model'arguments'c10

ModInteger'cpu'arguments'model'arguments'c11

ModInteger'cpu'arguments'model'arguments'c20

ModInteger'cpu'arguments'model'arguments'minimum

ModInteger'memory'arguments'intercept

ModInteger'memory'arguments'slope

MultiplyInteger'cpu'arguments'intercept

MultiplyInteger'cpu'arguments'slope

MultiplyInteger'memory'arguments'intercept

MultiplyInteger'memory'arguments'slope

NullList'cpu'arguments

NullList'memory'arguments

QuotientInteger'cpu'arguments'constant

QuotientInteger'cpu'arguments'model'arguments'c00

QuotientInteger'cpu'arguments'model'arguments'c01

QuotientInteger'cpu'arguments'model'arguments'c02

QuotientInteger'cpu'arguments'model'arguments'c10

QuotientInteger'cpu'arguments'model'arguments'c11

QuotientInteger'cpu'arguments'model'arguments'c20

QuotientInteger'cpu'arguments'model'arguments'minimum

QuotientInteger'memory'arguments'intercept

QuotientInteger'memory'arguments'minimum

QuotientInteger'memory'arguments'slope

RemainderInteger'cpu'arguments'constant

RemainderInteger'cpu'arguments'model'arguments'c00

RemainderInteger'cpu'arguments'model'arguments'c01

RemainderInteger'cpu'arguments'model'arguments'c02

RemainderInteger'cpu'arguments'model'arguments'c10

RemainderInteger'cpu'arguments'model'arguments'c11

RemainderInteger'cpu'arguments'model'arguments'c20

RemainderInteger'cpu'arguments'model'arguments'minimum

RemainderInteger'memory'arguments'intercept

RemainderInteger'memory'arguments'slope

SerialiseData'cpu'arguments'intercept

SerialiseData'cpu'arguments'slope

SerialiseData'memory'arguments'intercept

SerialiseData'memory'arguments'slope

Sha2_256'cpu'arguments'intercept

Sha2_256'cpu'arguments'slope

Sha2_256'memory'arguments

Sha3_256'cpu'arguments'intercept

Sha3_256'cpu'arguments'slope

Sha3_256'memory'arguments

SliceByteString'cpu'arguments'intercept

SliceByteString'cpu'arguments'slope

SliceByteString'memory'arguments'intercept

SliceByteString'memory'arguments'slope

SndPair'cpu'arguments

SndPair'memory'arguments

SubtractInteger'cpu'arguments'intercept

SubtractInteger'cpu'arguments'slope

SubtractInteger'memory'arguments'intercept

SubtractInteger'memory'arguments'slope

TailList'cpu'arguments

TailList'memory'arguments

Trace'cpu'arguments

Trace'memory'arguments

UnBData'cpu'arguments

UnBData'memory'arguments

UnConstrData'cpu'arguments

UnConstrData'memory'arguments

UnIData'cpu'arguments

UnIData'memory'arguments

UnListData'cpu'arguments

UnListData'memory'arguments

UnMapData'cpu'arguments

UnMapData'memory'arguments

VerifyEcdsaSecp256k1Signature'cpu'arguments

VerifyEcdsaSecp256k1Signature'memory'arguments

VerifyEd25519Signature'cpu'arguments'intercept

VerifyEd25519Signature'cpu'arguments'slope

VerifyEd25519Signature'memory'arguments

VerifySchnorrSecp256k1Signature'cpu'arguments'intercept

VerifySchnorrSecp256k1Signature'cpu'arguments'slope

VerifySchnorrSecp256k1Signature'memory'arguments

CekConstrCost'exBudgetCPU

CekConstrCost'exBudgetMemory

CekCaseCost'exBudgetCPU

CekCaseCost'exBudgetMemory

Bls12_381_G1_add'cpu'arguments

Bls12_381_G1_add'memory'arguments

Bls12_381_G1_compress'cpu'arguments

Bls12_381_G1_compress'memory'arguments

Bls12_381_G1_equal'cpu'arguments

Bls12_381_G1_equal'memory'arguments

Bls12_381_G1_hashToGroup'cpu'arguments'intercept

Bls12_381_G1_hashToGroup'cpu'arguments'slope

Bls12_381_G1_hashToGroup'memory'arguments

Bls12_381_G1_neg'cpu'arguments

Bls12_381_G1_neg'memory'arguments

Bls12_381_G1_scalarMul'cpu'arguments'intercept

Bls12_381_G1_scalarMul'cpu'arguments'slope

Bls12_381_G1_scalarMul'memory'arguments

Bls12_381_G1_uncompress'cpu'arguments

Bls12_381_G1_uncompress'memory'arguments

Bls12_381_G2_add'cpu'arguments

Bls12_381_G2_add'memory'arguments

Bls12_381_G2_compress'cpu'arguments

Bls12_381_G2_compress'memory'arguments

Bls12_381_G2_equal'cpu'arguments

Bls12_381_G2_equal'memory'arguments

Bls12_381_G2_hashToGroup'cpu'arguments'intercept

Bls12_381_G2_hashToGroup'cpu'arguments'slope

Bls12_381_G2_hashToGroup'memory'arguments

Bls12_381_G2_neg'cpu'arguments

Bls12_381_G2_neg'memory'arguments

Bls12_381_G2_scalarMul'cpu'arguments'intercept

Bls12_381_G2_scalarMul'cpu'arguments'slope

Bls12_381_G2_scalarMul'memory'arguments

Bls12_381_G2_uncompress'cpu'arguments

Bls12_381_G2_uncompress'memory'arguments

Bls12_381_finalVerify'cpu'arguments

Bls12_381_finalVerify'memory'arguments

Bls12_381_millerLoop'cpu'arguments

Bls12_381_millerLoop'memory'arguments

Bls12_381_mulMlResult'cpu'arguments

Bls12_381_mulMlResult'memory'arguments

Keccak_256'cpu'arguments'intercept

Keccak_256'cpu'arguments'slope

Keccak_256'memory'arguments

Blake2b_224'cpu'arguments'intercept

Blake2b_224'cpu'arguments'slope

Blake2b_224'memory'arguments

IntegerToByteString'cpu'arguments'c0

IntegerToByteString'cpu'arguments'c1

IntegerToByteString'cpu'arguments'c2

IntegerToByteString'memory'arguments'intercept

IntegerToByteString'memory'arguments'slope

ByteStringToInteger'cpu'arguments'c0

ByteStringToInteger'cpu'arguments'c1

ByteStringToInteger'cpu'arguments'c2

ByteStringToInteger'memory'arguments'intercept

ByteStringToInteger'memory'arguments'slope

AndByteString'cpu'arguments'intercept

AndByteString'cpu'arguments'slope1

AndByteString'cpu'arguments'slope2

AndByteString'memory'arguments'intercept

AndByteString'memory'arguments'slope

OrByteString'cpu'arguments'intercept

OrByteString'cpu'arguments'slope1

OrByteString'cpu'arguments'slope2

OrByteString'memory'arguments'intercept

OrByteString'memory'arguments'slope

XorByteString'cpu'arguments'intercept

XorByteString'cpu'arguments'slope1

XorByteString'cpu'arguments'slope2

XorByteString'memory'arguments'intercept

XorByteString'memory'arguments'slope

ComplementByteString'cpu'arguments'intercept

ComplementByteString'cpu'arguments'slope

ComplementByteString'memory'arguments'intercept

ComplementByteString'memory'arguments'slope

ReadBit'cpu'arguments

ReadBit'memory'arguments

WriteBits'cpu'arguments'intercept

WriteBits'cpu'arguments'slope

WriteBits'memory'arguments'intercept

WriteBits'memory'arguments'slope

ReplicateByte'cpu'arguments'intercept

ReplicateByte'cpu'arguments'slope

ReplicateByte'memory'arguments'intercept

ReplicateByte'memory'arguments'slope

ShiftByteString'cpu'arguments'intercept

ShiftByteString'cpu'arguments'slope

ShiftByteString'memory'arguments'intercept

ShiftByteString'memory'arguments'slope

RotateByteString'cpu'arguments'intercept

RotateByteString'cpu'arguments'slope

RotateByteString'memory'arguments'intercept

RotateByteString'memory'arguments'slope

CountSetBits'cpu'arguments'intercept

CountSetBits'cpu'arguments'slope

CountSetBits'memory'arguments

FindFirstSetBit'cpu'arguments'intercept

FindFirstSetBit'cpu'arguments'slope

FindFirstSetBit'memory'arguments

Ripemd_160'cpu'arguments'intercept

Ripemd_160'cpu'arguments'slope

Ripemd_160'memory'arguments

ExpModInteger'cpu'arguments'coefficient00

ExpModInteger'cpu'arguments'coefficient11

ExpModInteger'cpu'arguments'coefficient12

ExpModInteger'memory'arguments'intercept

ExpModInteger'memory'arguments'slope

DropList'cpu'arguments'intercept

DropList'cpu'arguments'slope

DropList'memory'arguments

LengthOfArray'cpu'arguments

LengthOfArray'memory'arguments

ListToArray'cpu'arguments'intercept

ListToArray'cpu'arguments'slope

ListToArray'memory'arguments'intercept

ListToArray'memory'arguments'slope

IndexArray'cpu'arguments

IndexArray'memory'arguments

Bls12_381_G1_multiScalarMul'cpu'arguments'intercept

Bls12_381_G1_multiScalarMul'cpu'arguments'slope

Bls12_381_G1_multiScalarMul'memory'arguments

Bls12_381_G2_multiScalarMul'cpu'arguments'intercept

Bls12_381_G2_multiScalarMul'cpu'arguments'slope

Bls12_381_G2_multiScalarMul'memory'arguments

InsertCoin'cpu'arguments'intercept

InsertCoin'cpu'arguments'slope

InsertCoin'memory'arguments'intercept

InsertCoin'memory'arguments'slope

LookupCoin'cpu'arguments'intercept

LookupCoin'cpu'arguments'slope

LookupCoin'memory'arguments

UnionValue'cpu'arguments'c00

UnionValue'cpu'arguments'c10

UnionValue'cpu'arguments'c01

UnionValue'cpu'arguments'c11

UnionValue'memory'arguments'intercept

UnionValue'memory'arguments'slope

ValueContains'cpu'arguments'constant

ValueContains'cpu'arguments'model'arguments'intercept

ValueContains'cpu'arguments'model'arguments'slope1

ValueContains'cpu'arguments'model'arguments'slope2

ValueContains'memory'arguments

ValueData'cpu'arguments'intercept

ValueData'cpu'arguments'slope

ValueData'memory'arguments'intercept

ValueData'memory'arguments'slope

UnValueData'cpu'arguments'c0

UnValueData'cpu'arguments'c1

UnValueData'cpu'arguments'c2

UnValueData'memory'arguments'intercept

UnValueData'memory'arguments'slope

ScaleValue'cpu'arguments'intercept

ScaleValue'cpu'arguments'slope

ScaleValue'memory'arguments'intercept

ScaleValue'memory'arguments'slope

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Context types

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextRedeemer :: Redeemer

Redeemer for the currently-executing script

scriptContextScriptInfo :: ScriptInfo

the purpose of the currently-executing script, along with information associated with the purpose

Bytes

newtype LedgerBytes

Constructors

LedgerBytes

Fields

getLedgerBytes :: BuiltinByteString

fromBytes :: ByteString -> LedgerBytes

Lift a Haskell bytestring to the Plutus abstraction LedgerBytes

Credentials

data StakingCredential

Staking credential used to assign rewards.

Constructors

StakingHash Credential
The staking hash is the Credential required to unlock a transaction output. Either a public key credential (PubKeyHash) or a script credential (ScriptHash). Both are hashed with BLAKE2b-244. 28 byte.

StakingPtr
The certificate pointer, constructed by the given slot number, transaction and certificate indices. NB: The fields should really be all Word64, as they are implemented in Word64, but RepIsStuckError StakingCredential :: Type -> Type) (Rep StakingCredential)))

data Credential

Credentials required to unlock a transaction output.

Constructors

PubKeyCredential PubKeyHash
The transaction that spends this output must be signed by the private key. See PubKeyHash.

ScriptCredential ScriptHash
The transaction that spends this output must include the validator script and be accepted by the validator. See ScriptHash.

Value

newtype Value

The Value type represents a collection of amounts of different currencies. We can think of Value as a vector space whose dimensions are currencies.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

There is no 'Ord Value' instance since Value is only a partial order, so compare can't do the right thing in some cases.

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Unroll Value = '[Value, CurrencySymbol, TokenName, Integer]

newtype CurrencySymbol

ByteString representing the currency, hashed with BLAKE2b-224. It is empty for Ada, 28 bytes for MintingPolicyHash. Forms an AssetClass along with TokenName. A Value is a map from CurrencySymbol's to a map from TokenName to an Integer.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

CurrencySymbol

Fields

unCurrencySymbol :: BuiltinByteString

newtype TokenName

ByteString of a name of a token. Shown as hex-encoded bytes. Should be no longer than 32 bytes, empty for Ada. Forms an AssetClass along with a CurrencySymbol.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TokenName

Fields

unTokenName :: BuiltinByteString

singleton :: CurrencySymbol -> TokenName -> Integer -> Value

Make a Value containing only the given quantity of the given currency.

unionWith :: (Integer -> Integer -> Integer) -> Value -> Value -> Value

Combine two Value maps with the argument function. Assumes the well-definedness of the two maps.

adaSymbol :: CurrencySymbol

The CurrencySymbol of the Ada currency.

adaToken :: TokenName

The TokenName of the Ada currency.

newtype Lovelace

Constructors

Lovelace

Fields

getLovelace :: Eq MintValue
Manual Eq instance: two MintValues are equal if they mint and burn the same assets, regardless of internal Map representation. Cannot use deriveEq for semantic equality.

Instance details

Defined in PlutusLedgerApi.V3.MintValue

Methods

(==) :: MintValue -> MintValue -> Unroll MintValue = '[MintValue, CurrencySymbol, TokenName, Integer]

emptyMintValue :: MintValue

mintValueToMap :: MintValue -> Map CurrencySymbol (Map TokenName Integer)

mintValueMinted :: MintValue -> Value

Get the Value minted by the MintValue.

mintValueBurned :: MintValue -> Value

Get the Value burned by the MintValue. All the negative quantities in the MintValue become positive in the resulting Value.

Time

newtype POSIXTime

POSIX time is measured as the number of milliseconds since 1970-01-01T00:00:00Z. This is not the same as Haskell's RepIsStuckError POSIXTime :: Type -> Type) (Rep POSIXTime)))

type POSIXTimeRange = Interval POSIXTime

An Interval of POSIXTimes.

Types for representing transactions

data Address

An address may contain two credentials, the payment credential and optionally a StakingCredential.

Constructors

Address

Fields

addressCredential :: Credential

the payment credential

addressStakingCredential :: RepIsStuckError Address :: Type -> Type) (Rep Address)))

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself. Hashed with BLAKE2b-224. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxInfo

TxInfo for PlutusV3

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

txInfoReferenceInputs :: [TxInInfo]

txInfoOutputs :: [TxOut]

txInfoFee :: Lovelace

txInfoMint :: MintValue

The Value minted by this transaction.

Invariant: This field does not contain Ada with zero quantity, unlike their namesakes in Plutus V1 and V2's ScriptContexts.

txInfoTxCerts :: [TxCert]

txInfoWdrl :: Map Credential Lovelace

txInfoValidRange :: POSIXTimeRange

txInfoSignatories :: [PubKeyHash]

txInfoRedeemers :: Map ScriptPurpose Redeemer

txInfoData :: Map DatumHash Datum

txInfoId :: TxId

txInfoVotes :: Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures :: [ProposalProcedure]

txInfoCurrentTreasuryAmount :: RepIsStuckError TxInfo :: Type -> Type) (Rep TxInfo)))

data TxOut

A transaction output, consisting of a target address, a value, optionally a datum/datum hash, and optionally a reference script.

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatum :: OutputDatum

txOutReferenceScript :: RepIsStuckError TxOut :: Type -> Type) (Rep TxOut)))

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

The transaction ID.

txOutRefIdx :: RepIsStuckError TxInInfo :: Type -> Type) (Rep TxInInfo)))

data OutputDatum

The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").

Constructors

NoOutputDatum

OutputDatumHash DatumHash

OutputDatum Datum

Intervals

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

The Eq instance gives equality of the intervals, not structural equality. There is no Enum in order to handle non-inclusive endpoints. For this reason, it may not be safe to use Intervals with non-inclusive endpoints on types whose Enum instances have partial methods.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

always :: Interval a

An Interval that covers every slot. In math. notation [-∞,+∞]

from :: a -> Interval a

from a is an Interval that includes all values that are greater than or equal to a. In math. notation: [a,+∞]

to :: a -> Interval a

to a is an Interval that includes all values that are smaller than or equal to a. In math. notation: [-∞,a]

lowerBound :: a -> LowerBound a

Construct a lower bound from a value. The resulting bound includes all values that are equal or greater than the input value.

upperBound :: a -> UpperBound a

Construct an upper bound from a value. The resulting bound includes all values that are equal or smaller than the input value.

strictLowerBound :: a -> LowerBound a

Construct a strict lower bound from a value. The resulting bound includes all values that are (strictly) greater than the input value.

strictUpperBound :: a -> UpperBound a

Construct a strict upper bound from a value. The resulting bound includes all values that are (strictly) smaller than the input value.

Ratio

data Rational

Represents an arbitrary-precision ratio.

The following two invariants are maintained:

The denominator is greater than zero.

The numerator and denominator are coprime.

ratio :: Integer -> Integer -> Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

unsafeRatio :: Integer -> Integer -> Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

numerator :: Rational -> Integer

Returns the numerator of its argument.

Note

It is not true in general that numerator <$> ratio x y = x; this will only hold if x and y are coprime. This is due to Rational normalizing the numerator and denominator.

denominator :: Rational -> Integer

Returns the denominator of its argument. This will always be greater than, or equal to, 1, although the type does not describe this.

Note

It is not true in general that denominator <$> ratio x y = y; this will only hold if x and y are coprime. This is due to Rational normalizing the numerator and denominator.

fromHaskellRatio :: Rational -> Rational

Converts a GHC Rational, preserving value.

Note: Does not work on-chain.

toHaskellRatio :: Rational -> Rational

Converts a Rational to a GHC Rational, preserving value.

Note: Does not work on-chain.

fromGHC :: Rational -> Rational

toGHC :: Rational -> Rational

Association maps

data Map k v

A Map of key-value pairs. A Map is considered well-defined if there are no key collisions, meaning that each value is uniquely identified by a key.

Use safeFromList to create well-defined Maps from arbitrary lists of pairs.

If cost minimisation is required, then you can use unsafeFromList but you must be certain that the list you are converting to a Map abides by the well-definedness condition.

Most operations on Maps are definedness-preserving, meaning that for the resulting Map to be well-defined then the input Map(s) have to also be well-defined. This is not checked explicitly unless mentioned in the documentation.

Take care when using fromBuiltinData and unsafeFromBuiltinData, as neither function performs deduplication of the input collection and may create invalid Maps!

(FromData k, FromData v) => FromData (Map k v)
A hand-written transformation from Data to Map. Compared to unsafeFromBuiltinData, it is safe to call when it is unknown if the Data is built with Datas Map constructor. Note that it is, however, unsafe in the sense that it assumes that any map encoded in the Data is well-formed, i.e. fromBuiltinData does not perform any deduplication of keys or of key-value pairs!

Instance details

Defined in PlutusTx.AssocMap

Methods

fromBuiltinData :: BuiltinData -> Maybe (Map k v)

(ToData k, ToData v) => ToData (Map k v)
Hand-written instances to use the underlying Map type in Data, and to be reasonably efficient.

Instance details

Defined in PlutusTx.AssocMap

Methods

toBuiltinData :: Map k v -> BuiltinData

(UnsafeFromData k, UnsafeFromData v) => UnsafeFromData (Map k v)
A hand-written transformation from Data to Map. It is unsafe because the caller must provide the guarantee that the Data is constructed using the Datas Map constructor. Note that it assumes, like the fromBuiltinData transformation, that the map encoded in the Data is well-formed, i.e. unsafeFromBuiltinData does not perform any deduplication of keys or of key-value pairs!

Instance details

Defined in Unroll (Map k v) = '[Map k v, k, v]

unsafeFromList :: [(k, v)] -> Map k v

Unsafely create a Map from a list of pairs. This should _only_ be applied to lists which have been checked to not contain duplicate keys, otherwise the resulting Map will contain conflicting entries (two entries sharing the same key). As usual, the "keys" are considered to be the first element of the pair.

Newtypes and hash types

newtype ScriptHash

Type representing the BLAKE2b-224 hash of a script. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

ScriptHash

Fields

getScriptHash :: BuiltinByteString

newtype RedeemerHash

Type representing the BLAKE2b-256 hash of a redeemer. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

RedeemerHash BuiltinByteString

newtype DatumHash

Type representing the BLAKE2b-256 hash of a datum. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

DatumHash BuiltinByteString

Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr DefaultUni Data)

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData ~Data

toData :: ToData a => a -> Data

Convert a value to Data.

Note: off-chain only.

fromData :: FromData a => Data -> Maybe a

Convert a value from Data, returning Nothing if this fails.

Note: off-chain only.

unsafeFromData :: UnsafeFromData a => Data -> a

Convert a value from Data, throwing if this fails.

Note: off-chain only.

dataToBuiltinData :: Data -> BuiltinData

NOT a builtin. Converts a Data into a BuiltinData. Only works off-chain.

builtinDataToData :: BuiltinData -> Data

NOT a builtin. Converts a BuiltinData into a Data. Only works off-chain.

Errors

class Monad m => MonadError e (m :: Type -> Type) | m -> e

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Minimal complete definition

BuiltinFailure logs. This is to compensate for the historical lack of error message content in operational errors (structural ones don't have this problem) in our evaluators (the CK and CEK machines). It would be better to fix the underlying issue and allow operational evaluation errors to carry some form of content, but for now we just fix the symptom in order for the end user to see the error message that they are supposed to see. The fix even makes some sense: what we do here is we emulate logging when the thrown unlifting error is an operational one, i.e. this is similar to what some builtins do manually (like when a crypto builtin fails and puts info about the failure into the logs).

Instance details

Defined in FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError !ScriptDecodeError
A deserialisation error TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

InvalidReturnValue
The script evaluated to a value that is not a valid return value.