PlutusLedgerApi.V3.Contexts

Documentation

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

newtype HotCommitteeCredential

Constructors

HotCommitteeCredential Credential

newtype DRepCredential

Constructors

DRepCredential Credential

data DRep

Constructors

DRep DRepCredential

DRepAlwaysAbstain

DRepAlwaysNoConfidence

data Delegatee

Constructors

DelegStake PubKeyHash

DelegVote DRep

DelegStakeVote PubKeyHash DRep

data TxCert

Constructors

TxCertRegStaking Credential (Maybe Value)
Register staking credential with an optional deposit amount

TxCertUnRegStaking Credential (Maybe Value)
Un-Register staking credential with an optional refund amount

TxCertDelegStaking Credential Delegatee
Delegate staking credential to a Delegatee

TxCertRegDeleg Credential Delegatee Value
Register and delegate staking credential to a Delegatee in one certificate. Noter that deposit is mandatory.

TxCertRegDRep DRepCredential Value
Register a DRep with a deposit value. The optional anchor is omitted.

TxCertUpdateDRep DRepCredential
Update a DRep. The optional anchor is omitted.

TxCertUnRegDRep DRepCredential Value
UnRegister a DRep with mandatory refund value

TxCertAuthHotCommittee ColdCommitteeCredential HotCommitteeCredential
Authorize a Hot credential for a specific Committee member's cold credential

TxCertResignColdCommittee ColdCommitteeCredential

data Voter

Constructors

CommitteeVoter HotCommitteeCredential

DRepVoter DRepCredential

StakePoolVoter PubKeyHash

data Vote

A vote. The optional anchor is omitted.

Constructors

VoteNo

VoteYes

Abstain

data GovernanceActionId

Similar to TxOutRef, but for GovActions

Constructors

GovernanceActionId

Fields

gaidTxId :: TxId

gaidGovActionIx :: Integer

data Committee

Constructors

Committee

Fields

committeeMembers :: Map ColdCommitteeCredential Integer

Committee members with epoch number when each of them expires

committeeQuorum :: Rational

Quorum of the committee that is necessary for a successful vote

newtype Constitution

A constitution. The optional anchor is omitted.

Constructors

Constitution

Fields

constitutionScript :: Maybe ScriptHash

data ProtocolVersion

Constructors

ProtocolVersion

Fields

pvMajor :: Integer

pvMinor :: Integer

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter name to the new value. Unchanged parameters are not included.

Constructors

ChangedParameters

Fields

getChangedParameters :: BuiltinData

data GovernanceAction

Constructors

ParameterChange (Maybe GovernanceActionId) ChangedParameters

HardForkInitiation (Maybe GovernanceActionId) ProtocolVersion
proposal to update protocol version

TreasuryWithdrawals (Map Credential Value)

NoConfidence (Maybe GovernanceActionId)

NewCommittee

Fields

(Maybe GovernanceActionId)

[ColdCommitteeCredential]

Old committee

Committee

New Committee

NewConstitution (Maybe GovernanceActionId) Constitution

InfoAction

data ProposalProcedure

A proposal procedure. The optional anchor is omitted.

Constructors

ProposalProcedure

Fields

ppDeposit :: Value

ppReturnAddr :: Credential

ppGovernanceAction :: GovernanceAction

data ScriptPurpose

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding Credential

Certifying TxCert

Voting Voter GovernanceActionId

Proposing Integer

data TxInfo

TxInfo for PlutusV3

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

txInfoReferenceInputs :: [TxInInfo]

txInfoOutputs :: [TxOut]

txInfoFee :: Value

txInfoMint :: Value

txInfoTxCerts :: [TxCert]

txInfoWdrl :: Map Credential Integer

txInfoValidRange :: POSIXTimeRange

txInfoSignatories :: [PubKeyHash]

txInfoRedeemers :: Map ScriptPurpose Redeemer

txInfoData :: Map DatumHash Datum

txInfoId :: TxId

txInfoVotes :: Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures :: [ProposalProcedure]

txInfoCurrentTreasuryAmount :: Maybe Value

txInfoTreasuryDonation :: Maybe Value

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextPurpose :: ScriptPurpose

the purpose of the currently-executing script