PlutusLedgerApi.V3.Data.Contexts

Documentation

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

newtype HotCommitteeCredential

Constructors

HotCommitteeCredential Credential

newtype DRepCredential

Constructors

DRepCredential Credential

data DRep

pattern DRep :: DRepCredential -> DRep

pattern DRepAlwaysAbstain :: DRep

pattern DRepAlwaysNoConfidence :: DRep

data Delegatee

pattern DelegStake :: PubKeyHash -> Delegatee

pattern DelegVote :: DRep -> Delegatee

pattern DelegStakeVote :: PubKeyHash -> DRep -> Delegatee

data TxCert

pattern TxCertRegStaking :: Credential -> Maybe Lovelace -> TxCert

pattern TxCertUnRegStaking :: Credential -> Maybe Lovelace -> TxCert

pattern TxCertDelegStaking :: Credential -> Delegatee -> TxCert

pattern TxCertRegDeleg :: Credential -> Delegatee -> Lovelace -> TxCert

pattern TxCertRegDRep :: DRepCredential -> Lovelace -> TxCert

pattern TxCertUpdateDRep :: DRepCredential -> TxCert

pattern TxCertUnRegDRep :: DRepCredential -> Lovelace -> TxCert

pattern TxCertPoolRegister :: PubKeyHash -> PubKeyHash -> TxCert

pattern TxCertPoolRetire :: PubKeyHash -> Integer -> TxCert

pattern TxCertAuthHotCommittee :: ColdCommitteeCredential -> HotCommitteeCredential -> TxCert

pattern TxCertResignColdCommittee :: ColdCommitteeCredential -> TxCert

data Voter

pattern CommitteeVoter :: HotCommitteeCredential -> Voter

pattern DRepVoter :: DRepCredential -> Voter

pattern StakePoolVoter :: PubKeyHash -> Voter

data Vote

A vote. The optional anchor is omitted.

pattern VoteNo :: Vote

pattern VoteYes :: Vote

pattern Abstain :: Vote

data GovernanceActionId

Similar to TxOutRef, but for GovActions

pattern GovernanceActionId :: TxId -> Integer -> GovernanceActionId

gaidTxId :: GovernanceActionId -> TxId

gaidGovActionIx :: GovernanceActionId -> Integer

data Committee

pattern Committee :: Map ColdCommitteeCredential Integer -> Rational -> Committee

committeeMembers :: Committee -> Map ColdCommitteeCredential Integer

committeeQuorum :: Committee -> Rational

newtype Constitution

A constitution. The optional anchor is omitted.

Constructors

Constitution

Fields

constitutionScript :: Maybe ScriptHash

data ProtocolVersion

pattern ProtocolVersion :: Integer -> Integer -> ProtocolVersion

pvMajor :: ProtocolVersion -> Integer

pvMinor :: ProtocolVersion -> Integer

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl. -- editorconfig-checker-disable-file

Invariant: This map is non-empty, and the keys are stored in ascending order.

Constructors

ChangedParameters

Fields

getChangedParameters :: BuiltinData

data GovernanceAction

pattern ParameterChange :: Maybe GovernanceActionId -> ChangedParameters -> Maybe ScriptHash -> GovernanceAction

pattern HardForkInitiation :: Maybe GovernanceActionId -> ProtocolVersion -> GovernanceAction

pattern TreasuryWithdrawals :: Map Credential Lovelace -> Maybe ScriptHash -> GovernanceAction

pattern NoConfidence :: Maybe GovernanceActionId -> GovernanceAction

pattern UpdateCommittee :: Maybe GovernanceActionId -> List ColdCommitteeCredential -> Map ColdCommitteeCredential Integer -> Rational -> GovernanceAction

pattern NewConstitution :: Maybe GovernanceActionId -> Constitution -> GovernanceAction

pattern InfoAction :: GovernanceAction

data ProposalProcedure

A proposal procedure. The optional anchor is omitted.

pattern ProposalProcedure :: Lovelace -> Credential -> GovernanceAction -> ProposalProcedure

ppDeposit :: ProposalProcedure -> Lovelace

ppReturnAddr :: ProposalProcedure -> Credential

ppGovernanceAction :: ProposalProcedure -> GovernanceAction

data ScriptPurpose

A ScriptPurpose uniquely identifies a Plutus script within a transaction.

pattern Minting :: CurrencySymbol -> ScriptPurpose

pattern Spending :: TxOutRef -> ScriptPurpose

pattern Rewarding :: Credential -> ScriptPurpose

pattern Certifying :: Integer -> TxCert -> ScriptPurpose

pattern Voting :: Voter -> ScriptPurpose

pattern Proposing :: Integer -> ProposalProcedure -> ScriptPurpose

data ScriptInfo

Like ScriptPurpose but with an optional datum for spending scripts.

pattern MintingScript :: CurrencySymbol -> ScriptInfo

pattern SpendingScript :: TxOutRef -> Maybe Datum -> ScriptInfo

pattern RewardingScript :: Credential -> ScriptInfo

pattern CertifyingScript :: Integer -> TxCert -> ScriptInfo

pattern VotingScript :: Voter -> ScriptInfo

pattern ProposingScript :: Integer -> ProposalProcedure -> ScriptInfo

data TxInInfo

An input of a pending transaction.

pattern TxInInfo :: TxOutRef -> TxOut -> TxInInfo

txInInfoOutRef :: TxInInfo -> TxOutRef

txInInfoResolved :: TxInInfo -> TxOut

data TxInfo

TxInfo for PlutusV3

pattern TxInfo :: List TxInInfo -> List TxInInfo -> List TxOut -> Lovelace -> Value -> List TxCert -> Map Credential Lovelace -> POSIXTimeRange -> List PubKeyHash -> Map ScriptPurpose Redeemer -> Map DatumHash Datum -> TxId -> Map Voter (Map GovernanceActionId Vote) -> List ProposalProcedure -> Maybe Lovelace -> Maybe Lovelace -> TxInfo

txInfoInputs :: TxInfo -> List TxInInfo

txInfoReferenceInputs :: TxInfo -> List TxInInfo

txInfoOutputs :: TxInfo -> List TxOut

txInfoFee :: TxInfo -> Lovelace

txInfoMint :: TxInfo -> Value

txInfoTxCerts :: TxInfo -> List TxCert

txInfoWdrl :: TxInfo -> Map Credential Lovelace

txInfoValidRange :: TxInfo -> POSIXTimeRange

txInfoSignatories :: TxInfo -> List PubKeyHash

txInfoRedeemers :: TxInfo -> Map ScriptPurpose Redeemer

txInfoData :: TxInfo -> Map DatumHash Datum

txInfoId :: TxInfo -> TxId

txInfoVotes :: TxInfo -> Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures :: TxInfo -> List ProposalProcedure

txInfoCurrentTreasuryAmount :: TxInfo -> Maybe Lovelace

txInfoTreasuryDonation :: TxInfo -> Maybe Lovelace

data ScriptContext

The context that the currently-executing script can access.

pattern ScriptContext :: TxInfo -> Redeemer -> ScriptInfo -> ScriptContext

scriptContextTxInfo :: ScriptContext -> TxInfo

scriptContextRedeemer :: ScriptContext -> Redeemer

scriptContextScriptInfo :: ScriptContext -> ScriptInfo

findOwnInput :: ScriptContext -> Maybe TxInInfo

Find the input currently being validated.

findDatum :: DatumHash -> TxInfo -> Maybe Datum

Find the data corresponding to a data hash, if there is one

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

Find the hash of a datum, if it is part of the pending transaction's hashes

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

Given a UTXO reference and a transaction (TxInfo), resolve it to one of the transaction's inputs (TxInInfo).

Note: this only searches the true transaction inputs and not the referenced transaction inputs.

findContinuingOutputs :: ScriptContext -> List Integer

Find the indices of all the outputs that pay to the same script address we are currently spending from, if any.

getContinuingOutputs :: ScriptContext -> List TxOut

Get all the outputs that pay to the same script address we are currently spending from, if any.

txSignedBy :: TxInfo -> PubKeyHash -> Bool

Check if a transaction was signed by the given public key.

Validator functions

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> List Value

Get the values paid to a public key address by a pending transaction.

valuePaidTo :: TxInfo -> PubKeyHash -> Value

Get the total value paid to a public key address by a pending transaction.

valueSpent :: TxInfo -> Value

Get the total value of inputs spent by this transaction.

valueProduced :: TxInfo -> Value

Get the total value of outputs produced by this transaction.

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

The CurrencySymbol of the current validator script.

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)