PlutusLedgerApi.V3.EvaluationContext

Documentation

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

mkEvaluationContext

Arguments

:: (MonadError CostModelApplyError m, MonadWriter [CostModelApplyWarn] m)

=> [Integer]
the (updated) cost model parameters of the protocol

-> m EvaluationContext

Build the EvaluationContext.

The input is a list of cost model parameters (which are integer values) passed from the ledger.

IMPORTANT: the cost model parameters MUST appear in the correct order, matching the names in ParamName. If the parameters are supplied in the wrong order then script cost calculations will be incorrect.

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

type CostModelParams = Map Text Integer

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

toMachineParameters :: MajorProtocolVersion -> EvaluationContext -> DefaultMachineParameters

data CostModelApplyError

Constructors

CMUnknownParamError !Text

CMInternalReadError

CMInternalWriteError !String

CMTooFewParamsError

Fields

cmTooFewExpected :: !Int

cmTooFewActual :: !Int