Plutus.V1.Ledger.Credential

Documentation

data StakingCredential

Staking credential used to assign rewards

Constructors

StakingHash Credential

StakingPtr Integer Integer Integer

data Credential

Credential required to unlock a transaction output

Constructors

PubKeyCredential PubKeyHash
The transaction that spends this output must be signed by the private key

ScriptCredential ValidatorHash
The transaction that spends this output must include the validator script and be accepted by the validator.