PlutusLedgerApi.Common

Description

The types and functions that are common among all ledger Plutus versions.

Script (de)serialization

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

serialiseCompiledCode :: forall a. CompiledCode a -> SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC :: Program DeBruijn DefaultUni DefaultFun () -> SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseUPLC :: SerialisedScript -> Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST.

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError DeserialiseFailure
an error from the underlying CBOR/serialise library

RemainderError ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang :: LedgerPlutusVersion

the script's plutus version

sdeIntroPv :: ProtocolVersion

the protocol version that will first introduce/enable the plutus version

sdeThisPv :: ProtocolVersion

the protocol version in which the error occurred

assertScriptWellFormed

Arguments

:: MonadError ScriptDecodeError m

=> LedgerPlutusVersion
the ledger Plutus version of the script.

-> ProtocolVersion
the current protocol version of the network

-> SerialisedScript
the script to check for well-formedness

-> m ()

Check if a Script is "valid" according to a protocol version. At the moment this means "deserialises correctly", which in particular implies that it is (almost certainly) an encoded script and the script does not mention any builtins unavailable in the given protocol version.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

Script evaluation

evaluateScriptCounting

Arguments

:: LedgerPlutusVersion
The ledger Plutus version of the script under execution.

-> ProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

evaluateScriptRestricting

Arguments

:: LedgerPlutusVersion
The ledger Plutus version of the script under execution.

-> ProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

data VerboseMode

A simple toggle indicating whether or not we should accumulate logs during script execution.

Constructors

Verbose
accumulate all traces

Quiet
don't accumulate anything

type LogOutput = [Text]

The type of the executed script's accumulated log output: a list of Text.

It will be an empty list if the VerboseMode is set to Quiet.

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError ScriptDecodeError
A deserialisation error

IncompatibleVersionError (Version ())
An error indicating a version tag that we don't support TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

Network's versioning

The network's behaviour (and plutus's by extension) can change via hard forks, which directly correspond to major-number protocol version bumps.

data ProtocolVersion

This represents the Cardano protocol version, with its major and minor components. This relies on careful understanding between us and the ledger as to what this means.

Constructors

ProtocolVersion

Fields

pvMajor :: Int

the major component

pvMinor :: Int

the minor component

data LedgerPlutusVersion

The ledger Plutus versions. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the AST's Version

Constructors

PlutusV1
introduced in shelley era

PlutusV2
introduced in vasil era

PlutusV3
not yet enabled, probably will be introduced in chang era

builtinsIntroducedIn :: Map (LedgerPlutusVersion, ProtocolVersion) (Set DefaultFun)

A map indicating which builtin functions were introduced in which ProtocolVersion. Each builtin function should appear at most once.

This must be updated when new builtins are added. See Note [New builtins and protocol versions]

builtinsAvailableIn :: LedgerPlutusVersion -> ProtocolVersion -> Set DefaultFun

Which builtin functions are available in the given ProtocolVersion?

See Note [New builtins and protocol versions]

Network's costing parameters

A less drastic approach (that does not rely on a HF) to affect the network's (and plutus's by extension) behaviour is by tweaking the values of the cost model parameters.

The network does not associate names to cost model parameters; Plutus attaches names to the network's cost model parameters (values) either in a raw textual form or typed by a specific plutus version.

See Note [Cost model parameters]

type CostModelParams = Map Text Integer

A raw representation of the ledger's cost model parameters.

The associated keys/names to the parameter values are arbitrarily set by the plutus team; the ledger does not hold any such names.

See Note [Cost model parameters]

toCostModelParams :: IsParamName k => [(k, Integer)] -> CostModelParams

Untags the plutus version from the typed cost model parameters and returns their raw textual form (internally used by CostModelInterface).

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

class IsParamName a

A parameter name for different plutus versions.

Each Plutus version should expose such an enumeration as an ADT and create an instance of ParamName out of it.

A valid parameter name has to be enumeration, bounded, ordered, and prettyprintable to a "lower-Kebab" string.

Minimal complete definition

showParamName

showParamName :: IsParamName a => a -> String

Take the raw textual form for a given typed-by-plutus-version cost model parameter

data GenericParamName a

A Generic wrapper for use with deriving via

Evaluation context

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

There are two sets of parameters: one is with immediate unlifting and the other one is with deferred unlifting. We have to keep both of them, because depending on the language version either one has to be used or the other. We also compile them separately due to all the inlining and optimization that need to happen for things to be efficient.

Constructors

EvaluationContext

Fields

machineParametersImmediate :: DefaultMachineParameters

machineParametersDeferred :: DefaultMachineParameters

mkDynEvaluationContext :: MonadError CostModelApplyError m => BuiltinVersion DefaultFun -> CostModelParams -> m EvaluationContext

Create an EvaluationContext for a given plutus-builtins' version.

The input is a Map of Texts to cost integer values (aka CostModelParams, CostModel) See Note [Inlining meanings of builtins].

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

toMachineParameters :: ProtocolVersion -> EvaluationContext -> DefaultMachineParameters

Get the raw machine parameters (as understood by the CEK evaluator) stored inside an EvaluationContext.