PlutusLedgerApi.Common

Description

The types and functions that are common among all ledger Plutus versions.

Script (de)serialization

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

data ScriptForEvaluation

A Plutus script ready to be evaluated on-chain, via evaluateScriptRestricting.

serialisedScript :: ScriptForEvaluation -> SerialisedScript

Get a SerialisedScript from a ScriptForEvaluation. O(1).

deserialisedScript :: ScriptForEvaluation -> ScriptNamedDeBruijn

Get a ScriptNamedDeBruijn from a ScriptForEvaluation. O(1).

serialiseCompiledCode :: forall a. CompiledCode a -> SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC :: Program DeBruijn DefaultUni DefaultFun () -> SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseScript

Arguments

:: forall m. MonadError ScriptDecodeError m

=> PlutusLedgerLanguage
the Plutus ledger language of the script.

-> MajorProtocolVersion
which major protocol version the script was submitted in.

-> SerialisedScript
the script to deserialise.

-> m ScriptForEvaluation

The deserialization from a serialised script into a ScriptForEvaluation, ready to be evaluated on-chain. Called inside phase-1 validation (i.e., deserialisation error is a phase-1 error).

uncheckedDeserialiseUPLC :: SerialisedScript -> Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST. Does *not* do ledger-language-version-specific checks like for allowable builtins.

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError !DeserialiseFailureInfo
an error from the underlying CBOR/serialise library

RemainderError !ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LedgerLanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang :: !PlutusLedgerLanguage

the script's ledger language

sdeIntroPv :: !MajorProtocolVersion

the major protocol version that will first introduce/enable the ledger language

sdeThisPv :: !MajorProtocolVersion

the current protocol version

PlutusCoreLanguageNotAvailableError

Fields

sdeAffectedVersion :: !Version

the Plutus Core language of the script under execution.

sdeThisLang :: !PlutusLedgerLanguage

the Plutus ledger language of the script under execution.

sdeThisPv :: !MajorProtocolVersion

the current protocol version

newtype ScriptNamedDeBruijn

A script with named de-bruijn indices.

Constructors

ScriptNamedDeBruijn (Program NamedDeBruijn DefaultUni DefaultFun ())

Script evaluation

evaluateScriptCounting

Arguments

:: PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

-> MajorProtocolVersion
Which major protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ScriptForEvaluation
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

evaluateScriptRestricting

Arguments

:: PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

-> MajorProtocolVersion
Which major protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> ScriptForEvaluation
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

Note: Parameterized over the LedgerPlutusVersion since 1. The builtins allowed (during decoding) differ, and 2. The Plutus language versions allowed differ.

evaluateTerm :: ExBudgetMode cost DefaultUni DefaultFun -> MajorProtocolVersion -> VerboseMode -> EvaluationContext -> Term NamedDeBruijn DefaultUni DefaultFun () -> (Either (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun) (Term NamedDeBruijn DefaultUni DefaultFun ()), cost, [Text])

Evaluate a fully-applied term using the CEK machine. Useful for mimicking the behaviour of the on-chain evaluator.

data VerboseMode

A simple toggle indicating whether or not we should accumulate logs during script execution.

Constructors

Verbose
accumulate all traces

Quiet
don't accumulate anything

type LogOutput = [Text]

The type of the executed script's accumulated log output: a list of Text.

It will be an empty list if the VerboseMode is set to Quiet.

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError !(CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError !FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError !ScriptDecodeError
A deserialisation error TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

data Data

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

Network's versioning

The network's behaviour (and plutus's by extension) can change via hard forks, which directly correspond to major-number protocol version bumps.

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion :: Int

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in shelley era

PlutusV2
introduced in vasil era

PlutusV3
not yet enabled

data Version

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

builtinsIntroducedIn :: Map (PlutusLedgerLanguage, MajorProtocolVersion) (Set DefaultFun)

A map indicating which builtin functions were introduced in which MajorProtocolVersion. Each builtin function should appear at most once.

This must be updated when new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn :: PlutusLedgerLanguage -> MajorProtocolVersion -> Set DefaultFun

Which builtin functions are available in the given given PlutusLedgerLanguage and MajorProtocolVersion?

See Note [New builtins/language versions and protocol versions]

ledgerLanguageIntroducedIn :: PlutusLedgerLanguage -> MajorProtocolVersion

Query the protocol version that a specific Plutus ledger language was first introduced in.

ledgerLanguagesAvailableIn :: MajorProtocolVersion -> Set PlutusLedgerLanguage

Which Plutus language versions are available in the given MajorProtocolVersion?

See Note [New builtins/language versions and protocol versions]

Network's costing parameters

A less drastic approach (that does not rely on a HF) to affect the network's (and plutus's by extension) behaviour is by tweaking the values of the cost model parameters.

The network does not associate names to cost model parameters; Plutus attaches names to the network's cost model parameters (values) either in a raw textual form or typed by a specific plutus version.

See Note [Cost model parameters]

type CostModelParams = Map Text Integer

toCostModelParams :: IsParamName p => [(p, Integer)] -> CostModelParams

Untags the plutus version from the typed cost model parameters and returns their raw textual form (internally used by CostModelInterface).

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

class (Enum a, Bounded a) => IsParamName a where

A parameter name for different plutus versions.

Each Plutus version should expose such an enumeration as an ADT and create an instance of ParamName out of it.

A valid parameter name has to be enumeration, bounded, ordered, and prettyprintable to a "lower-Kebab" string.

Minimal complete definition

showParamName

Methods

showParamName :: a -> Text

Produce the raw textual form for a given typed-by-plutus-version cost model parameter Any implementation *must be* an injective function. The GIsParamName generic implementation guarantees injectivity.

readParamName :: Text -> Maybe a

default implementation that inverts the showParamName operation (not very efficient)

data GenericParamName a

A Generic wrapper for use with deriving via

data CostModelApplyError

Constructors

CMUnknownParamError !Text

CMInternalReadError

CMInternalWriteError !String

CMTooFewParamsError

Fields

cmTooFewExpected :: !Int

cmTooFewActual :: !Int

data CostModelApplyWarn

Constructors

CMTooManyParamsWarn

Fields

cmTooManyExpected :: !Int

cmTooManyActual :: !Int

Evaluation context

newtype EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

Constructors

EvaluationContext

Fields

machineParameters :: DefaultMachineParameters

mkDynEvaluationContext :: MonadError CostModelApplyError m => BuiltinSemanticsVariant DefaultFun -> CostModelParams -> m EvaluationContext

Create an EvaluationContext for a given builtin semantics variant.

The input is a Map of Texts to cost integer values (aka CostModelParams, CostModel) See Note [Inlining meanings of builtins].

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

toMachineParameters :: MajorProtocolVersion -> EvaluationContext -> DefaultMachineParameters

mkTermToEvaluate

Arguments

:: MonadError EvaluationError m

=> PlutusLedgerLanguage
the Plutus ledger language of the script under execution.

-> MajorProtocolVersion
which major protocol version to run the operation in

-> ScriptForEvaluation
the script to evaluate

-> [Data]
the arguments that the script's underlying term will be applied to

-> m (Term NamedDeBruijn DefaultUni DefaultFun ())

Shared helper for the evaluation functions: evaluateScriptCounting and evaluateScriptRestricting,

Given a ScriptForEvaluation:

1) applies the term to a list of Data arguments (e.g. Datum, Redeemer, ScriptContext) 2) checks that the applied-term is well-scoped 3) returns the applied-term