PlutusLedgerApi.Common

Description

The types and functions that are common among all ledger Plutus versions.

Script (de)serialization

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

serialiseCompiledCode :: forall a. CompiledCode a -> SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC :: Program DeBruijn DefaultUni DefaultFun () -> SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseUPLC :: SerialisedScript -> Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST.

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError !DeserialiseFailure
an error from the underlying CBOR/serialise library

RemainderError !ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang :: !PlutusLedgerLanguage

the script's plutus version

sdeIntroPv :: !ProtocolVersion

the protocol version that will first introduce/enable the plutus version

sdeThisPv :: !ProtocolVersion

the protocol version in which the error occurred

assertScriptWellFormed

Arguments

:: MonadError ScriptDecodeError m

=> PlutusLedgerLanguage
the ledger Plutus version of the script.

-> ProtocolVersion
the current protocol version of the network

-> SerialisedScript
the script to check for well-formedness

-> m ()

Check if a Script is "valid" according to a protocol version. At the moment this means "deserialises correctly", which in particular implies that it is (almost certainly) an encoded script and the script does not mention any builtins unavailable in the given protocol version.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

Script evaluation

evaluateScriptCounting

Arguments

:: PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

-> ProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

evaluateScriptRestricting

Arguments

:: PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

-> ProtocolVersion
Which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

Note: Parameterized over the LedgerPlutusVersion since 1. The builtins allowed (during decoding) differ, and 2. The Plutus language versions allowed differ.

data VerboseMode

A simple toggle indicating whether or not we should accumulate logs during script execution.

Constructors

Verbose
accumulate all traces

Quiet
don't accumulate anything

type LogOutput = [Text]

The type of the executed script's accumulated log output: a list of Text.

It will be an empty list if the VerboseMode is set to Quiet.

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError !(CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError !FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError !ScriptDecodeError
A deserialisation error

IncompatibleVersionError !Version
An error indicating a version tag that we don't support TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

Network's versioning

The network's behaviour (and plutus's by extension) can change via hard forks, which directly correspond to major-number protocol version bumps.

data ProtocolVersion

This represents the Cardano protocol version, with its major and minor components. This relies on careful understanding between us and the ledger as to what this means.

Constructors

ProtocolVersion

Fields

pvMajor :: Int

the major component

pvMinor :: Int

the minor component

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in shelley era

PlutusV2
introduced in vasil era

PlutusV3
not yet enabled

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

builtinsIntroducedIn :: Map (PlutusLedgerLanguage, ProtocolVersion) (Set DefaultFun)

A map indicating which builtin functions were introduced in which ProtocolVersion. Each builtin function should appear at most once.

This must be updated when new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn :: PlutusLedgerLanguage -> ProtocolVersion -> Set DefaultFun

Which builtin functions are available in the given given PlutusLedgerLanguage and ProtocolVersion?

See Note [New builtins/language versions and protocol versions]

Network's costing parameters

A less drastic approach (that does not rely on a HF) to affect the network's (and plutus's by extension) behaviour is by tweaking the values of the cost model parameters.

The network does not associate names to cost model parameters; Plutus attaches names to the network's cost model parameters (values) either in a raw textual form or typed by a specific plutus version.

See Note [Cost model parameters]

type CostModelParams = Map Text Integer

A raw representation of the ledger's cost model parameters.

The associated keys/names to the parameter values are arbitrarily set by the plutus team; the ledger does not hold any such names.

See Note [Cost model parameters]

toCostModelParams :: IsParamName k => [(k, Integer)] -> CostModelParams

Untags the plutus version from the typed cost model parameters and returns their raw textual form (internally used by CostModelInterface).

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

class IsParamName a

A parameter name for different plutus versions.

Each Plutus version should expose such an enumeration as an ADT and create an instance of ParamName out of it.

A valid parameter name has to be enumeration, bounded, ordered, and prettyprintable to a "lower-Kebab" string.

Minimal complete definition

showParamName

showParamName :: IsParamName a => a -> String

Take the raw textual form for a given typed-by-plutus-version cost model parameter

data GenericParamName a

A Generic wrapper for use with deriving via

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError !Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError !String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

CMTooFewParamsError
See Note [Cost model parameters from the ledger's point of view]

Fields

cmTooFewExpected :: !Int

cmTooFewActual :: !Int

data CostModelApplyWarn

A non-fatal warning when trying to create a cost given some plain costmodel parameters.

Constructors

CMTooManyParamsWarn
More costmodel parameters given, than expected

See Note [Cost model parameters from the ledger's point of view]

Fields

cmTooManyExpected :: !Int

cmTooManyActual :: !Int

Evaluation context

newtype EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

Constructors

EvaluationContext

Fields

machineParameters :: DefaultMachineParameters

mkDynEvaluationContext :: MonadError CostModelApplyError m => BuiltinVersion DefaultFun -> CostModelParams -> m EvaluationContext

Create an EvaluationContext for a given plutus-builtins' version.

The input is a Map of Texts to cost integer values (aka CostModelParams, CostModel) See Note [Inlining meanings of builtins].

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

toMachineParameters :: ProtocolVersion -> EvaluationContext -> DefaultMachineParameters

mkTermToEvaluate

Arguments

:: MonadError EvaluationError m

=> PlutusLedgerLanguage
the Plutus ledger language of the script under execution.

-> ProtocolVersion
which protocol version to run the operation in

-> SerialisedScript
the script to evaluate in serialised form

-> [Data]
the arguments that the script's underlying term will be applied to

-> m (Term NamedDeBruijn DefaultUni DefaultFun ())

Shared helper for the evaluation functions: evaluateScriptCounting and evaluateScriptRestricting,

Given a SerialisedScript:

1) deserialises to a plutus Term and checks its ast version 2) applies the term to a list of Data arguments (e.g. Datum, Redeemer, ScriptContext) 3) checks that the applied-term is well-scoped 4) returns the applied-term