PlutusLedgerApi.Common.Versions

Description

This module contains the protocol versions and plutus versions as witnessed by the cardano ledger.

Cardano Protocol versions

data ProtocolVersion

This represents the Cardano protocol version, with its major and minor components. This relies on careful understanding between us and the ledger as to what this means.

Constructors

ProtocolVersion

Fields

pvMajor :: Int

the major component

pvMinor :: Int

the minor component

Protocol Version aliases

Based on https://github.com/input-output-hk/cardano-ledger/wiki/First-Block-of-Each-Era

shelleyPV :: ProtocolVersion

Shelley era was introduced in protocol version 2.0

allegraPV :: ProtocolVersion

Allegra era was introduced in protocol version 3.0

maryPV :: ProtocolVersion

Mary era was introduced in protocol version 4.0

alonzoPV :: ProtocolVersion

Alonzo era was introduced in protocol version 5.0

vasilPV :: ProtocolVersion

Vasil era was introduced in protocol version 7.0

changPV :: ProtocolVersion

Chang era is not yet published. It will probably be introduced in protocol version 8.0

Plutus versions

data LedgerPlutusVersion

The ledger Plutus versions. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the AST's Version

Constructors

PlutusV1
introduced in shelley era

PlutusV2
introduced in vasil era

PlutusV3
not yet enabled, probably will be introduced in chang era

Version-testing functions

languageIntroducedIn :: LedgerPlutusVersion -> ProtocolVersion

Query the protocol version that a specific ledger Plutus version was first introduced in.

Introduction in this context means the enablement/allowance of scripts of that ledger Plutus version to be executed on-chain.

languagesAvailableIn :: ProtocolVersion -> Set LedgerPlutusVersion

Given a protocol version return a set of all available ledger Plutus versions that are enabled/allowed to run.

Assumes that the ledger Plutus versions once introduced/enabled, will never be disabled in the future.

builtinsIntroducedIn :: Map (LedgerPlutusVersion, ProtocolVersion) (Set DefaultFun)

A map indicating which builtin functions were introduced in which ProtocolVersion. Each builtin function should appear at most once.

This must be updated when new builtins are added. See Note [New builtins and protocol versions]

builtinsAvailableIn :: LedgerPlutusVersion -> ProtocolVersion -> Set DefaultFun

Which builtin functions are available in the given ProtocolVersion?

See Note [New builtins and protocol versions]