PlutusLedgerApi.Common.Versions

Description

This module contains the code for handling the various kinds of version that we care about:

Protocol versions

Plutus ledger languages

Plutus Core language versions

Cardano Protocol versions

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion :: Doc ann

Decoder s [MajorProtocolVersion]

type Rep MajorProtocolVersion

Instance details

Defined in PlutusLedgerApi.Common.ProtocolVersions

type Rep MajorProtocolVersion = D1 ('MetaData "MajorProtocolVersion" "PlutusLedgerApi.Common.ProtocolVersions" "plutus-ledger-api-1.60.0.0-inplace" 'True) (C1 ('MetaCons "MajorProtocolVersion" 'PrefixI 'True) (S1 ('MetaSel ('Just "getMajorProtocolVersion") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Int)))

Protocol Version aliases

Based on https://github.com/IntersectMBO/cardano-ledger/wiki/First-Block-of-Each-Era

shelleyPV :: MajorProtocolVersion

Shelley era was introduced in protocol version 2.0

allegraPV :: MajorProtocolVersion

Allegra era was introduced in protocol version 3.0

maryPV :: MajorProtocolVersion

Mary era was introduced in protocol version 4.0

alonzoPV :: MajorProtocolVersion

Alonzo era was introduced in protocol version 5.0

vasilPV :: MajorProtocolVersion

The Vasil HF introduced the Babbage era and Plutus V2

valentinePV :: MajorProtocolVersion

Valentine was an intra-era HF where builtin functions VerifyEcdsaSecp256k1Signature and VerifySchnorrSecp256k1Signature were enabled.

changPV :: MajorProtocolVersion

The Chang HF introduced the Conway era and Plutus V3

plominPV :: MajorProtocolVersion

The Plomin HF was an intra-era HF where some new builtin functions were introduced in Plutus V2 and V3.

vanRossemPV :: MajorProtocolVersion

The van Rossem HF will be an intra-era HF.

newestPV :: MajorProtocolVersion

knownPVs :: [MajorProtocolVersion]

The set of protocol versions that are "known", i.e. that have been released and have actual differences associated with them. This is currently only used for testing, so efficiency is not parmount and a list is fine.

futurePV :: MajorProtocolVersion

This is a placeholder for when we don't yet know what protocol version will be used for something. It's a very high protocol version that should never appear in reality.

We should not assign names to future protocol versions until it's confirmed that they are correct, otherwise we could accidentally associate something with the wrong protocol version.

Plutus ledger languages

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in Alonzo HF

PlutusV2
introduced in Vasil HF

PlutusV3
introduced in Chang HF

Decoder s [PlutusLedgerLanguage]

type Rep PlutusLedgerLanguage

Instance details

Defined in PlutusLedgerApi.Common.Versions

type Rep PlutusLedgerLanguage = D1 ('MetaData "PlutusLedgerLanguage" "PlutusLedgerApi.Common.Versions" "plutus-ledger-api-1.60.0.0-inplace" 'False) (C1 ('MetaCons "PlutusV1" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "PlutusV2" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "PlutusV3" 'PrefixI 'False) (U1 :: Type -> Type)))

Plutus Core language versions

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor :: PlutusCore.Version
type Rep Version = D1 ('MetaData "Version" "PlutusCore.Version" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "Version" 'PrefixI 'True) (S1 ('MetaSel ('Just "_versionMajor") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Natural) :*: (S1 ('MetaSel ('Just "_versionMinor") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Natural) :*: S1 ('MetaSel ('Just "_versionPatch") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Natural))))

Version-testing functions

ledgerLanguageIntroducedIn :: PlutusLedgerLanguage -> MajorProtocolVersion

Query the protocol version that a specific Plutus ledger language was first introduced in.

ledgerLanguagesAvailableIn :: MajorProtocolVersion -> Set PlutusLedgerLanguage

Which Plutus language versions are available in the given MajorProtocolVersion? See Note [New builtins/language versions and protocol versions]. This function (and others in this module) assumes that once a LL is available it remains available in all later PVs and that if m <= n, PlutusVm is introduced no later than PlutusVn.

plcVersionsIntroducedIn :: PlutusLedgerLanguage -> Map MajorProtocolVersion (Set Version)

A map indicating which Plutus Core versions were introduced in which MajorProtocolVersion and PlutusLedgerLanguage. Each version should appear at most once. This must be updated when new versions are added. See Note [New builtins/language versions and protocol versions]

plcVersionsAvailableIn :: PlutusLedgerLanguage -> MajorProtocolVersion -> Set Version

Which Plutus Core language versions are available in the given PlutusLedgerLanguage and MajorProtocolVersion?

builtinsIntroducedIn :: PlutusLedgerLanguage -> DefaultFun)

Given a ledger language, return a map indicating which builtin functions were introduced in which MajorProtocolVersion. This must be updated when new builtins are added. It is not necessary to add entries for protocol versions where no new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn :: PlutusLedgerLanguage -> MajorProtocolVersion -> DefaultFun

Return a set containing the builtins which are available in a given LL in a given PV. All builtins are available in all LLs from vanRossemPV onwards.

batch1 :: [DefaultFun]