PlutusLedgerApi.V1

Description

The interface to Plutus V1 for the ledger.

Scripts

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

serialiseCompiledCode :: forall a. CompiledCode a -> SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC :: Program DeBruijn DefaultUni DefaultFun () -> SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseUPLC :: SerialisedScript -> Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST.

Validating scripts

assertScriptWellFormed

Arguments

:: MonadError ScriptDecodeError m

=> ProtocolVersion
which protocol version to run the operation in

-> SerialisedScript
the script to check for well-formedness

-> m ()

Check if a Script is "valid" according to a protocol version. At the moment this means "deserialises correctly", which in particular implies that it is (almost certainly) an encoded script and the script does not mention any builtins unavailable in the given protocol version.

Running scripts

evaluateScriptRestricting

Arguments

:: ProtocolVersion
which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> ExBudget
The resource budget which must not be exceeded during evaluation

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

evaluateScriptCounting

Arguments

:: ProtocolVersion
which protocol version to run the operation in

-> VerboseMode
Whether to produce log output

-> EvaluationContext
Includes the cost model to use for tallying up the execution costs

-> SerialisedScript
The script to evaluate

-> [Data]
The arguments to the script

-> (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Protocol version

data ProtocolVersion

This represents the Cardano protocol version, with its major and minor components. This relies on careful understanding between us and the ledger as to what this means.

Constructors

ProtocolVersion

Fields

pvMajor :: Int

the major component

pvMinor :: Int

the minor component

Verbose mode and log output

data VerboseMode

A simple toggle indicating whether or not we should accumulate logs during script execution.

Constructors

Verbose
accumulate all traces

Quiet
don't accumulate anything

type LogOutput = [Text]

The type of the executed script's accumulated log output: a list of Text.

It will be an empty list if the VerboseMode is set to Quiet.

Costing-related types

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

newtype ExCPU

Counts CPU units in picoseconds: maximum value for SatInt is 2^63 ps, or appproximately 106 days.

Constructors

ExCPU CostingInteger

newtype ExMemory

Counts size in machine words.

Constructors

ExMemory CostingInteger

data SatInt

Cost model

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

There are two sets of parameters: one is with immediate unlifting and the other one is with deferred unlifting. We have to keep both of them, because depending on the language version either one has to be used or the other. We also compile them separately due to all the inlining and optimization that need to happen for things to be efficient.

mkEvaluationContext

Arguments

:: (MonadError CostModelApplyError m, MonadWriter [CostModelApplyWarn] m)

=> [Integer]
the (updated) cost model parameters of the protocol

-> m EvaluationContext

Build the EvaluationContext.

The input is a list of integer values passed from the ledger and are expected to appear in correct order.

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

data ParamName

The enumeration of all possible cost model parameter names for this language version.

IMPORTANT: The order of appearance of the data constructors here matters. DO NOT REORDER. See Note [Quotation marks in cost model parameter constructors] See Note [Cost model parameters from the ledger's point of view]

Constructors

AddInteger'cpu'arguments'intercept

AddInteger'cpu'arguments'slope

AddInteger'memory'arguments'intercept

AddInteger'memory'arguments'slope

AppendByteString'cpu'arguments'intercept

AppendByteString'cpu'arguments'slope

AppendByteString'memory'arguments'intercept

AppendByteString'memory'arguments'slope

AppendString'cpu'arguments'intercept

AppendString'cpu'arguments'slope

AppendString'memory'arguments'intercept

AppendString'memory'arguments'slope

BData'cpu'arguments

BData'memory'arguments

Blake2b_256'cpu'arguments'intercept

Blake2b_256'cpu'arguments'slope

Blake2b_256'memory'arguments

CekApplyCost'exBudgetCPU

CekApplyCost'exBudgetMemory

CekBuiltinCost'exBudgetCPU

CekBuiltinCost'exBudgetMemory

CekConstCost'exBudgetCPU

CekConstCost'exBudgetMemory

CekDelayCost'exBudgetCPU

CekDelayCost'exBudgetMemory

CekForceCost'exBudgetCPU

CekForceCost'exBudgetMemory

CekLamCost'exBudgetCPU

CekLamCost'exBudgetMemory

CekStartupCost'exBudgetCPU

CekStartupCost'exBudgetMemory

CekVarCost'exBudgetCPU

CekVarCost'exBudgetMemory

ChooseData'cpu'arguments

ChooseData'memory'arguments

ChooseList'cpu'arguments

ChooseList'memory'arguments

ChooseUnit'cpu'arguments

ChooseUnit'memory'arguments

ConsByteString'cpu'arguments'intercept

ConsByteString'cpu'arguments'slope

ConsByteString'memory'arguments'intercept

ConsByteString'memory'arguments'slope

ConstrData'cpu'arguments

ConstrData'memory'arguments

DecodeUtf8'cpu'arguments'intercept

DecodeUtf8'cpu'arguments'slope

DecodeUtf8'memory'arguments'intercept

DecodeUtf8'memory'arguments'slope

DivideInteger'cpu'arguments'constant

DivideInteger'cpu'arguments'model'arguments'intercept

DivideInteger'cpu'arguments'model'arguments'slope

DivideInteger'memory'arguments'intercept

DivideInteger'memory'arguments'minimum

DivideInteger'memory'arguments'slope

EncodeUtf8'cpu'arguments'intercept

EncodeUtf8'cpu'arguments'slope

EncodeUtf8'memory'arguments'intercept

EncodeUtf8'memory'arguments'slope

EqualsByteString'cpu'arguments'constant

EqualsByteString'cpu'arguments'intercept

EqualsByteString'cpu'arguments'slope

EqualsByteString'memory'arguments

EqualsData'cpu'arguments'intercept

EqualsData'cpu'arguments'slope

EqualsData'memory'arguments

EqualsInteger'cpu'arguments'intercept

EqualsInteger'cpu'arguments'slope

EqualsInteger'memory'arguments

EqualsString'cpu'arguments'constant

EqualsString'cpu'arguments'intercept

EqualsString'cpu'arguments'slope

EqualsString'memory'arguments

FstPair'cpu'arguments

FstPair'memory'arguments

HeadList'cpu'arguments

HeadList'memory'arguments

IData'cpu'arguments

IData'memory'arguments

IfThenElse'cpu'arguments

IfThenElse'memory'arguments

IndexByteString'cpu'arguments

IndexByteString'memory'arguments

LengthOfByteString'cpu'arguments

LengthOfByteString'memory'arguments

LessThanByteString'cpu'arguments'intercept

LessThanByteString'cpu'arguments'slope

LessThanByteString'memory'arguments

LessThanEqualsByteString'cpu'arguments'intercept

LessThanEqualsByteString'cpu'arguments'slope

LessThanEqualsByteString'memory'arguments

LessThanEqualsInteger'cpu'arguments'intercept

LessThanEqualsInteger'cpu'arguments'slope

LessThanEqualsInteger'memory'arguments

LessThanInteger'cpu'arguments'intercept

LessThanInteger'cpu'arguments'slope

LessThanInteger'memory'arguments

ListData'cpu'arguments

ListData'memory'arguments

MapData'cpu'arguments

MapData'memory'arguments

MkCons'cpu'arguments

MkCons'memory'arguments

MkNilData'cpu'arguments

MkNilData'memory'arguments

MkNilPairData'cpu'arguments

MkNilPairData'memory'arguments

MkPairData'cpu'arguments

MkPairData'memory'arguments

ModInteger'cpu'arguments'constant

ModInteger'cpu'arguments'model'arguments'intercept

ModInteger'cpu'arguments'model'arguments'slope

ModInteger'memory'arguments'intercept

ModInteger'memory'arguments'minimum

ModInteger'memory'arguments'slope

MultiplyInteger'cpu'arguments'intercept

MultiplyInteger'cpu'arguments'slope

MultiplyInteger'memory'arguments'intercept

MultiplyInteger'memory'arguments'slope

NullList'cpu'arguments

NullList'memory'arguments

QuotientInteger'cpu'arguments'constant

QuotientInteger'cpu'arguments'model'arguments'intercept

QuotientInteger'cpu'arguments'model'arguments'slope

QuotientInteger'memory'arguments'intercept

QuotientInteger'memory'arguments'minimum

QuotientInteger'memory'arguments'slope

RemainderInteger'cpu'arguments'constant

RemainderInteger'cpu'arguments'model'arguments'intercept

RemainderInteger'cpu'arguments'model'arguments'slope

RemainderInteger'memory'arguments'intercept

RemainderInteger'memory'arguments'minimum

RemainderInteger'memory'arguments'slope

Sha2_256'cpu'arguments'intercept

Sha2_256'cpu'arguments'slope

Sha2_256'memory'arguments

Sha3_256'cpu'arguments'intercept

Sha3_256'cpu'arguments'slope

Sha3_256'memory'arguments

SliceByteString'cpu'arguments'intercept

SliceByteString'cpu'arguments'slope

SliceByteString'memory'arguments'intercept

SliceByteString'memory'arguments'slope

SndPair'cpu'arguments

SndPair'memory'arguments

SubtractInteger'cpu'arguments'intercept

SubtractInteger'cpu'arguments'slope

SubtractInteger'memory'arguments'intercept

SubtractInteger'memory'arguments'slope

TailList'cpu'arguments

TailList'memory'arguments

Trace'cpu'arguments

Trace'memory'arguments

UnBData'cpu'arguments

UnBData'memory'arguments

UnConstrData'cpu'arguments

UnConstrData'memory'arguments

UnIData'cpu'arguments

UnIData'memory'arguments

UnListData'cpu'arguments

UnListData'memory'arguments

UnMapData'cpu'arguments

UnMapData'memory'arguments

VerifyEd25519Signature'cpu'arguments'intercept

VerifyEd25519Signature'cpu'arguments'slope

VerifyEd25519Signature'memory'arguments

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

CMTooFewParamsError
See Note [Cost model parameters from the ledger's point of view]

Fields

cmTooFewExpected :: Int

cmTooFewActual :: Int

type CostModelParams = Map Text Integer

A raw representation of the ledger's cost model parameters.

The associated keys/names to the parameter values are arbitrarily set by the plutus team; the ledger does not hold any such names.

See Note [Cost model parameters]

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Context types

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextPurpose :: ScriptPurpose

the purpose of the currently-executing script

data ScriptPurpose

Purpose of the script that is currently running

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

Supporting types used in the context types

ByteStrings

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

toBuiltin :: ToBuiltin a arep => a -> arep

fromBuiltin :: FromBuiltin arep a => arep -> a

Bytes

newtype LedgerBytes

Constructors

LedgerBytes

Fields

getLedgerBytes :: BuiltinByteString

fromBytes :: ByteString -> LedgerBytes

Lift a Haskell bytestring to the Plutus abstraction LedgerBytes

Certificates

data DCert

A representation of the ledger DCert. Some information is digested, and not included

Constructors

DCertDelegRegKey StakingCredential

DCertDelegDeRegKey StakingCredential

DCertDelegDelegate

Fields

StakingCredential

delegator

PubKeyHash

delegatee

DCertPoolRegister
A digest of the PoolParams

Fields

PubKeyHash

poolId

PubKeyHash

pool VFR

DCertPoolRetire PubKeyHash Integer
The retirement certificate and the Epoch in which the retirement will take place

DCertGenesis
A really terse Digest

DCertMir
Another really terse Digest

Credentials

data StakingCredential

Staking credential used to assign rewards.

Constructors

StakingHash Credential
The staking hash is the Credential required to unlock a transaction output. Either a public key credential (PubKeyHash) or a script credential (ScriptHash). Both are hashed with BLAKE2b-244. 28 byte.

StakingPtr
The certificate pointer, constructed by the given slot number, transaction and certificate indices. NB: The fields should really be all Word64, as they are implemented in Word64, but Integer is our only integral type so we need to use it instead.

Fields

Integer

the slot number

Integer

the transaction index (within the block)

Integer

the certificate index (within the transaction)

data Credential

Credentials required to unlock a transaction output.

Constructors

PubKeyCredential PubKeyHash
The transaction that spends this output must be signed by the private key. See PubKeyHash.

ScriptCredential ScriptHash
The transaction that spends this output must include the validator script and be accepted by the validator. See ScriptHash.

Value

newtype Value

The Value type represents a collection of amounts of different currencies. We can think of Value as a vector space whose dimensions are currencies. To create a value of Value, we need to specify a currency. This can be done using adaValueOf. To get the ada dimension of Value we use fromValue. Plutus contract authors will be able to define modules similar to Ada for their own currencies.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

There is no 'Ord Value' instance since Value is only a partial order, so compare can't do the right thing in some cases.

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Integer)

newtype CurrencySymbol

ByteString representing the currency, hashed with BLAKE2b-224. It is empty for Ada, 28 bytes for MintingPolicyHash. Forms an AssetClass along with TokenName. A Value is a map from CurrencySymbol's to a map from TokenName to an Integer.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

CurrencySymbol

Fields

unCurrencySymbol :: BuiltinByteString

newtype TokenName

ByteString of a name of a token. Shown as UTF-8 string when possible. Should be no longer than 32 bytes, empty for Ada. Forms an AssetClass along with a CurrencySymbol.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TokenName

Fields

unTokenName :: BuiltinByteString

singleton :: CurrencySymbol -> TokenName -> Integer -> Value

Make a Value containing only the given quantity of the given currency.

unionWith :: (Integer -> Integer -> Integer) -> Value -> Value -> Value

adaSymbol :: CurrencySymbol

The CurrencySymbol of the Ada currency.

adaToken :: TokenName

The TokenName of the Ada currency.

Time

newtype POSIXTime

POSIX time is measured as the number of milliseconds since 1970-01-01T00:00:00Z. This is not the same as Haskell's POSIXTime

Constructors

POSIXTime

Fields

getPOSIXTime :: Integer

type POSIXTimeRange = Interval POSIXTime

An Interval of POSIXTimes.

Types for representing transactions

data Address

An address may contain two credentials, the payment credential and optionally a StakingCredential.

Constructors

Address

Fields

addressCredential :: Credential

the payment credential

addressStakingCredential :: Maybe StakingCredential

the staking credential

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself. Hashed with BLAKE2b-224. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxInfo

A pending transaction. This is the view as seen by validator scripts, so some details are stripped out.

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

Transaction inputs; cannot be an empty list

txInfoOutputs :: [TxOut]

Transaction outputs

txInfoFee :: Value

The fee paid by this transaction.

txInfoMint :: Value

The Value minted by this transaction.

txInfoDCert :: [DCert]

Digests of certificates included in this transaction

txInfoWdrl :: [(StakingCredential, Integer)]

Withdrawals

txInfoValidRange :: POSIXTimeRange

The valid range for the transaction.

txInfoSignatories :: [PubKeyHash]

Signatures provided with the transaction, attested that they all signed the tx

txInfoData :: [(DatumHash, Datum)]

The lookup table of datums attached to the transaction

txInfoId :: TxId

Hash of the pending transaction body (i.e. transaction excluding witnesses)

data TxOut

A transaction output, consisting of a target address (Address), a value (Value), and optionally a datum hash (DatumHash).

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

The transaction ID.

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

Intervals

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

The Equality of two intervals is specified as the canonical, structural equality and not the equality of the elements of their two underlying sets; the same holds for Ord.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data Extended a

A set extended with a positive and negative infinity.

Constructors

NegInf

Finite a

PosInf

type Closure = Bool

Whether a bound is inclusive or not.

data UpperBound a

The upper bound of an interval.

Constructors

UpperBound (Extended a) Closure

data LowerBound a

The lower bound of an interval.

Constructors

LowerBound (Extended a) Closure

always :: Interval a

An Interval that covers every slot. In math. notation [-∞,+∞]

from :: a -> Interval a

from a is an Interval that includes all values that are greater than or equal to a. In math. notation: [a,+∞]

to :: a -> Interval a

to a is an Interval that includes all values that are smaller than or equal to a. In math. notation: [-∞,a]

lowerBound :: a -> LowerBound a

Construct a lower bound from a value.

The resulting bound includes all values that are equal or greater than the input value.

upperBound :: a -> UpperBound a

Construct an upper bound from a value.

The resulting bound includes all values that are equal or smaller than the input value.

strictLowerBound :: a -> LowerBound a

Construct a strict lower bound from a value.

The resulting bound includes all values that are (strictly) greater than the input value.

strictUpperBound :: a -> UpperBound a

Construct a strict upper bound from a value.

The resulting bound includes all values that are (strictly) smaller than the input value.

Newtypes and hash types

newtype ScriptHash

Script runtime representation of a Digest SHA256.

Constructors

ScriptHash

Fields

getScriptHash :: BuiltinByteString

newtype Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

Constructors

Redeemer

Fields

getRedeemer :: BuiltinData

newtype RedeemerHash

Type representing the BLAKE2b-256 hash of a redeemer. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

RedeemerHash BuiltinByteString

newtype Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

Constructors

Datum

Fields

getDatum :: BuiltinData

newtype DatumHash

Script runtime representation of a Digest SHA256.

Constructors

DatumHash BuiltinByteString

Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData Data

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

class FromData a where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData a where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

toData :: ToData a => a -> Data

Convert a value to Data.

fromData :: FromData a => Data -> Maybe a

Convert a value from Data, returning Nothing if this fails.

dataToBuiltinData :: Data -> BuiltinData

Convert a Data into a BuiltinData. Only works off-chain.

builtinDataToData :: BuiltinData -> Data

Convert a BuiltinData into a Data. Only works off-chain.

Errors

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError ScriptDecodeError
A deserialisation error

IncompatibleVersionError (Version ())
An error indicating a version tag that we don't support TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError DeserialiseFailure
an error from the underlying CBOR/serialise library

RemainderError ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang :: LedgerPlutusVersion

the script's plutus version

sdeIntroPv :: ProtocolVersion

the protocol version that will first introduce/enable the plutus version

sdeThisPv :: ProtocolVersion

the protocol version in which the error occurred