PlutusLedgerApi.V3.Contexts

data TxCert

Constructors

TxCertRegStaking Credential (Maybe Lovelace)
Register staking credential with an optional deposit amount

TxCertUnRegStaking Credential (Maybe Lovelace)
Un-Register staking credential with an optional refund amount

TxCertDelegStaking Credential Delegatee
Delegate staking credential to a Delegatee

TxCertRegDeleg Credential Delegatee Lovelace
Register and delegate staking credential to a Delegatee in one certificate. Note that deposit is mandatory.

TxCertRegDRep DRepCredential Lovelace
Register a DRep with a deposit value. The optional anchor is omitted.

TxCertUpdateDRep DRepCredential
Update a DRep. The optional anchor is omitted.

TxCertUnRegDRep DRepCredential Lovelace
UnRegister a DRep with mandatory refund value

TxCertPoolRegister
A digest of the PoolParams

Fields

PubKeyHash

poolId

PubKeyHash

pool VFR

TxCertPoolRetire PubKeyHash RepIsStuckError ProtocolVersion :: Type -> Type) (Rep ProtocolVersion)))

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl.

Invariant: This map is non-empty, and the keys are stored in ascending order.

This Data object has the following format (in pseudocode):

ChangedParametersData = Map ChangedIdData ChangedManyValueData ChangedIdData = I Integer ChangedManyValueData = ChangedSingleValueData | List[ChangedSingleValueData...] -- ^ an arbitrary-length, heterogeneous (integer or ratio) list of values (to support sub-parameters)

ChangedSingleValueData = I Integer -- a proposed integer value | List[I Integer, I Integer] -- a proposed numerator,denominator (ratio value) -- ^ a 2-exact element list; *BE CAREFUL* because this can be alternatively (ambiguously) interpreted -- as a many-value data (sub-parameter) of two integer single-value data.

, where Map,I,List are the constructors of Data and Integer is the usual arbitrary-precision PlutusTx/Haskell Integer.

Constructors

ChangedParameters

Fields

getChangedParameters :: BuiltinData

data TxInfo

TxInfo for PlutusV3

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

txInfoReferenceInputs :: [TxInInfo]

txInfoOutputs :: [TxOut]

txInfoFee :: Lovelace

txInfoMint :: MintValue

The Value minted by this transaction.

Invariant: This field does not contain Ada with zero quantity, unlike their namesakes in Plutus V1 and V2's ScriptContexts.

txInfoTxCerts :: [TxCert]

txInfoWdrl :: Map Credential Lovelace

txInfoValidRange :: POSIXTimeRange

txInfoSignatories :: [PubKeyHash]

txInfoRedeemers :: Map ScriptPurpose Redeemer

txInfoData :: Map DatumHash Datum

txInfoId :: TxId

txInfoVotes :: Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures :: [ProposalProcedure]

txInfoCurrentTreasuryAmount :: RepIsStuckError TxInfo :: Type -> Type) (Rep TxInfo)))

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextRedeemer :: Redeemer

Redeemer for the currently-executing script

scriptContextScriptInfo :: ScriptInfo

the purpose of the currently-executing script, along with information associated with the purpose

findOwnInput :: ScriptContext -> Maybe TxInInfo

Find the input currently being validated.

findDatum :: DatumHash -> TxInfo -> Maybe Datum

Find the data corresponding to a data hash, if there is one

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

Find the hash of a datum, if it is part of the pending transaction's hashes

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

Given a UTXO reference and a transaction (TxInfo), resolve it to one of the transaction's inputs (TxInInfo).

Note: this only searches the true transaction inputs and not the referenced transaction inputs.

findContinuingOutputs :: ScriptContext -> [Integer]

Find the indices of all the outputs that pay to the same script address we are currently spending from, if any.

getContinuingOutputs :: ScriptContext -> [TxOut]

Get all the outputs that pay to the same script address we are currently spending from, if any.

txSignedBy :: TxInfo -> PubKeyHash -> Bool

Check if a transaction was signed by the given public key.

Validator functions

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> [Value]

Get the values paid to a public key address by a pending transaction.

valuePaidTo :: TxInfo -> PubKeyHash -> Value

Get the total value paid to a public key address by a pending transaction.

valueSpent :: TxInfo -> Value

Get the total value of inputs spent by this transaction.

valueProduced :: TxInfo -> Value

Get the total value of outputs produced by this transaction.

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

The CurrencySymbol of the current validator script.

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)